% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{analysis1}
\alias{analysis1}
\title{analysis1 data set}
\format{
A data frame with 4680 rows and 36 variables
}
\usage{
analysis1
}
\description{
This is the complete example analysis data set.
}
\seealso{
\link{pmt_first}, \link{pmt_pk}, \link{pmt_obs}
}
\keyword{datasets}
