% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption.R
\name{as.caption}
\alias{as.caption}
\title{Create caption}
\usage{
as.caption(
  x,
  short = NULL,
  short_repeat = TRUE,
  short_sep = NULL,
  write = FALSE
)
}
\arguments{
\item{x}{caption text.}

\item{short}{an abbreviated form of the caption to be used in a list of
tables.}

\item{short_repeat}{logical; if a short caption is provided, it will be
repeated at the start of the main caption.}

\item{short_sep}{a character sequence used to separate the short title
with the main caption when a short title is specified.}

\item{write}{logical; if \code{TRUE}, caption will be written to output file
when \code{\link[=stable_save]{stable_save()}} is called; this argument does not apply to
\code{stable_long} objects whose captions are \emph{always} written to output the
output file.}
}
\value{
An object inheriting from \code{st_caption} and \code{character}.
}
\description{
Create caption
}
\details{
Caption text (in \code{x}) can contain a short caption title enclosed in brackets
(\verb{[]}) as the opening characters (see examples). The short title is
assumed to start after the opening \code{[} and end prior to \verb{]} when all brackets
are matched, allowing \verb{[...]} in the short title as long as all brackets
are matched. For other scenarios where a single \verb{]} may need to appear
in the short title, the user should use the \code{short} argument instead of
the bracket notation.
}
\examples{

# No separate short title
as.caption("Main title text")

# Note that the short title is repeated by default
as.caption("[Short title]. Main title text")

unclass(as.caption("[Short title]. Main title text."))
unclass(as.caption(". Main title text.", short = "Short"))

}
