% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{as.list.pt_opts}
\alias{as.list.pt_opts}
\title{Coerce pt_opts object to a list}
\usage{
\method{as.list}{pt_opts}(x, ...)
}
\arguments{
\item{x}{the \code{pt_opts} object}

\item{...}{not used}
}
\description{
Coerce pt_opts object to a list
}
\examples{
x <- as.list(pt)

}
