% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{as_glossary}
\alias{as_glossary}
\title{Create a new glossary object}
\usage{
as_glossary(...)
}
\arguments{
\item{...}{unquoted \verb{<label> = <definition>} pairs, or a named list
of abbreviations where the names are the labels.}
}
\value{
A new glossary object.
}
\description{
Create a new glossary object
}
\details{
For all entries in the new object, the abbreviation is set to the label.
See \code{\link[=update_abbrev]{update_abbrev()}} to customize the abbreviation.

Entries that have duplicate labels will be dropped with a warning.
}
\examples{

as_glossary(wt = "weight", ht = "height", bmi = "body mass index")

as_glossary(wt = "weight", ht = "height", wt = "baseline weight")

l <- list(VPC = "visual predictive check", tz = "timezone")

as_glossary(l)

as_glossary(l, wt = "weight")

}
\seealso{
\code{\link[=read_glossary]{read_glossary()}}, \code{\link[=update_abbrev]{update_abbrev()}}

\code{\link[=update_abbrev]{update_abbrev()}}
}
