% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-stable.R
\name{as_stable}
\alias{as_stable}
\alias{as_stable.pmtable}
\alias{as_stable.stable}
\title{Create stable from pmtable}
\usage{
as_stable(x, ...)

\method{as_stable}{pmtable}(x, ..., wrap = FALSE, wrapw = FALSE)

\method{as_stable}{stable}(x, ...)
}
\arguments{
\item{x}{object to convert to stable}

\item{...}{for the \code{pmtable} method, these are extra named arguments to pass
to \code{\link[=stable]{stable()}}}

\item{wrap}{if \code{TRUE}, the stable output will be wrapped in a latex table
environment}

\item{wrapw}{if \code{TRUE}, the stable output will be wrapped in a latex table
environment and
the output will be written to \code{\link[=stdout]{stdout()}}; use this along with
\code{results = "asis"} when rendering tables with \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\description{
Create stable from pmtable
}
\keyword{internal}
