% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{$<-.pt_opts}
\alias{$<-.pt_opts}
\title{Assign a pmtables option}
\usage{
\method{$}{pt_opts}(x, i) <- value
}
\arguments{
\item{x}{the \code{pt_opts} object}

\item{i}{the name of the option to set}

\item{value}{the value to set}
}
\description{
Assign a pmtables option
}
\examples{

pt_opts$bq_col <- "BLQ"

}
