% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{cat_data}
\alias{cat_data}
\title{Summarize categorical data}
\usage{
cat_data(
  data,
  cols,
  by = ".total",
  panel = by,
  summarize_all = TRUE,
  all_name = "All",
  wide = FALSE,
  nby = NULL,
  complete = FALSE,
  denom = c("group", "total")
)
}
\arguments{
\item{data}{the data frame to summarize; the user should filter or subset
so that \code{data} contains exactly the records to be summarized; pmtables will
not add or remove rows prior to summarizing \code{data}}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{by}{a grouping variable; may be character vector or quosure}

\item{panel}{data set column name to stratify the summary}

\item{summarize_all}{logical indicating whether or not to include a summary
of the full data in the output.}

\item{all_name}{label for full data summary.}

\item{wide}{\code{logical}; if \code{TRUE}, data frame will be returned in wide format;
if \code{FALSE}, it will be returned in \code{long} format.}

\item{nby}{number of unique levels for the \code{by} variable.}

\item{complete}{logial; if \code{TRUE}, then data the summary will be completed
for missing levels of \code{by}and \code{panel}.}

\item{denom}{the denominator to use when calculating percent for each level;
\code{group} uses the total number in the chunk being summarized; \code{total} uses
the total number in the data set; historically, \code{group} has been used, but
\code{total} may be more what the reader is expecting.}
}
\description{
Summarize categorical data
}
\examples{

cat_data(pmt_first, cols = c(SEX = "SEXf", RF = "RFf"), by = "STUDYf")

}
