% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics-table.R
\name{check_sum_func}
\alias{check_sum_func}
\title{Check if summary function evaluates to the correct format}
\usage{
check_sum_func(sum_func)
}
\arguments{
\item{sum_func}{The summary function to use for summarizing the continuous data in \code{\link[=pt_demographics]{pt_demographics()}}}
}
\description{
Check if summary function evaluates to the correct format
}
\keyword{internal}
