% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pmtable.R
\name{class-pmtable}
\alias{class-pmtable}
\title{Class: pmtable}
\description{
Objects with class \code{pmtable} are return from the pmtable-provided data
summary functions:
\itemize{
\item \code{\link[=pt_cont_wide]{pt_cont_wide()}}
\item \code{\link[=pt_cont_long]{pt_cont_long()}}
\item \code{\link[=pt_cat_wide]{pt_cat_wide()}}
\item \code{\link[=pt_cat_long]{pt_cat_long()}}
\item \code{\link[=pt_data_inventory]{pt_data_inventory()}}
}
}
\details{
The object is a list, with an item called \code{data}, which is the summarized
data set, as well as other items that are arguments for \code{\link[=stable]{stable()}} or
\code{\link[=stable_long]{stable_long()}}.
}
\examples{
tab <- pt_cont_long(pmt_first, cols = "WT,AGE,SCR")

str(tab)

methods(class = "pmtable")

}
