% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-align.R
\name{col_ragged}
\alias{col_ragged}
\alias{col_fixed}
\title{Format a table cell with fixed width}
\usage{
col_ragged(
  size,
  unit = "cm",
  ragged = c("right", "left", "no"),
  coltype = c("p", "m", "b")
)

col_fixed(..., ragged = "no")
}
\arguments{
\item{size}{the width of the cell}

\item{unit}{to go with \code{size}}

\item{ragged}{use \code{right} to left justify and \code{left} to right justify}

\item{coltype}{column type}
}
\description{
Format a table cell with fixed width
}
