% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-span.R
\name{col_spanners}
\alias{col_spanners}
\title{Create groups of columns with spanners}
\usage{
col_spanners(span = NULL, span_split = NULL, cols = NUL)
}
\arguments{
\item{span}{a list of objects created with \code{\link[=colgroup]{colgroup()}}; ; see also \code{\link[=st_span]{st_span()}}}

\item{span_split}{not implemented at this time; ; see also \code{\link[=st_span_split]{st_span_split()}}}
}
\description{
Create groups of columns with spanners
}
