% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-align.R
\name{cols_align}
\alias{cols_align}
\alias{cols_center}
\alias{cols_left}
\alias{cols_right}
\title{Set table column alignment}
\usage{
cols_align(
  .default = c("l", "c", "r"),
  ...,
  .rt = NULL,
  .cent = NULL,
  .lt = NULL,
  .complete = NULL
)

cols_center(...)

cols_left(...)

cols_right(...)
}
\arguments{
\item{.default}{the default column alignment}

\item{...}{exceptions to use from \code{default}; each argument should be named
by a column in the data set; values should be either "l", "c", or "r"}

\item{.rt}{column names as character vector or comma separated string
to right justify}

\item{.cent}{column names as character vector or comma separated string
to center}

\item{.lt}{column names as character vector or comma separated string
to right justify}

\item{.complete}{not used}
}
\description{
Set table column alignment
}
