% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-functions.R
\name{cont_long_fun}
\alias{cont_long_fun}
\title{Function for continuous long summaries}
\usage{
cont_long_fun(
  value,
  digit_fun = sig,
  id = NULL,
  digits = 3,
  na_fill = "--",
  ...
)
}
\arguments{
\item{value}{the data to summarize}

\item{digit_fun}{a function to format digits in the summaries}

\item{id}{a vector of subject IDs; same length as \code{value}}

\item{digits}{the number of digits in the summary; the current implementation
passes \code{digits} to \code{digit_fun()}}

\item{na_fill}{value to fill with when all values in the summary are missing}

\item{...}{not used}
}
\value{
The function returns a tibble with one row and five columns:
\enumerate{
\item \code{n} the number of non-missing observations in \code{value}
\item \code{Mean} the mean of \code{value}
\item \code{Median} the median of \code{value}
\item \code{SD} the standard deviation of \code{value}
\item \code{Min / Max} the range of value
}

All columns are returned as \code{character}.
}
\description{
Function for continuous long summaries
}
\examples{
pmtables:::cont_long_fun(rnorm(100))

}
\keyword{internal}
