% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{data_inventory_chunk}
\alias{data_inventory_chunk}
\title{Create a data inventory summary for a data chunk}
\usage{
data_inventory_chunk(
  data,
  by,
  panel = by,
  stacked = FALSE,
  tot = FALSE,
  all_name = "all",
  dv_col = "DV",
  bq_col = "BQL",
  id_col = "ID",
  ...
)
}
\arguments{
\item{data}{the data frame to summarize; the user should filter or subset
so that \code{data} contains exactly the records to be summarized; pmtables will
not add or remove rows prior to summarizing \code{data}}

\item{by}{The outer grouping variable; may be character or quosure.}

\item{panel}{The panel grouping variable; may be character or quosure.}

\item{stacked}{If \code{TRUE}, then independent summaries are created by \code{outer}
and included in a single table (see examples).}

\item{tot}{logical indicating if a summary row should be included}

\item{all_name}{a name to use for the complete data summary}

\item{dv_col}{Character name of \code{DV} column.}

\item{bq_col}{Character name of \code{BQL} column; see \code{\link[=find_bq_col]{find_bq_col()}}.}

\item{id_col}{Character name of \code{ID} column.}

\item{...}{used to absorb other arguments; not used}
}
\description{
Create a data inventory summary for a data chunk
}
