% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-clear_reps.R
\name{deduplicate_values}
\alias{deduplicate_values}
\title{Clear grouped column values}
\usage{
deduplicate_values(data, cols = groups(data), extra_groups = NULL)
}
\arguments{
\item{data}{data frame}

\item{cols}{columns to clear}

\item{extra_groups}{extra grouping variables}
}
\description{
Clear grouped column values
}
