% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-utils.R
\name{df_grep_rows}
\alias{df_grep_rows}
\alias{df_grepl_rows}
\title{Find matching rows in a data frame}
\usage{
df_grep_rows(data, pattern, cols = names(data))

df_grepl_rows(data, pattern, cols = names(data))
}
\arguments{
\item{data}{a data frame to search}

\item{pattern}{a regular expression}

\item{cols}{a character vector of column names to search}
}
\description{
Find matching rows in a data frame
}
