% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_path}
\alias{format_path}
\title{Format the path to an image file}
\usage{
format_path(file, dir = NULL, path.type = c("proj", "none", "raw"))
}
\arguments{
\item{file}{the name of the image file.}

\item{dir}{the directory where the image file is stored.}

\item{path.type}{a character string indicating how the path to the image
file should be formatted; use \code{"proj"} to have the path expressed relative
to the RStudio project file; use \code{"none"} to format without any directory
information (just the image file name); use \code{"raw"} to print the complete
path to the file using \code{dir} as-is.}
}
\value{
A string with the formatted image file path.
}
\description{
Format the path to an image file
}
\details{
Note that the default value for \code{path.type} is \code{"proj"}. This requires
that an RStudio project file is able to be found using
\code{\link[rprojroot:find_root]{rprojroot::find_root()}} with the \link[rprojroot:criteria]{rprojroot::is_rstudio_project} criterion.

Once mrggsave finds a root for a given working directory, it caches the value
for the remainder of the R session. An error will be generated if an image is
attempted to be saved using \code{path.type="proj"} but an RStudio project file
was not able to be located.
}
\examples{
\dontrun{
format_path("foo.txt", "my/path")
format_path("foo.txt", "my/path", path.type = "none")
format_path("foo.txt", "my/path", path.type = "raw")
}

}
