% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-table.R
\name{get_debug_data}
\alias{get_debug_data}
\title{Get debug information from stable object}
\usage{
get_debug_data(x)
}
\arguments{
\item{x}{an stable object}
}
\description{
Get debug information from stable object
}
