% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-stable.R
\name{get_stable_data}
\alias{get_stable_data}
\title{Get debug information from stable object}
\usage{
get_stable_data(x)
}
\arguments{
\item{x}{an stable object}
}
\description{
Get debug information from stable object
}
