% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{glossary_notes}
\alias{glossary_notes}
\alias{glossary_notes.character}
\alias{glossary_notes.list}
\alias{glossary_notes.glossary}
\title{Return formatted notes from a tex glossary file or glossary object}
\usage{
glossary_notes(x, ...)

\method{glossary_notes}{character}(x, ..., format = guess_glo_fmt(x))

\method{glossary_notes}{list}(x, ...)

\method{glossary_notes}{glossary}(x, ..., sep = ": ", collapse = "; ")
}
\arguments{
\item{x}{path to a tex glossary file, a glossary object, or a list that can
be coerced to a glossary object with \code{\link[=as_glossary]{as_glossary()}}.}

\item{...}{unquoted glossary labels to be included in the notes text.}

\item{format}{passed to \code{\link[=read_glossary]{read_glossary()}}.}

\item{sep}{character to separate name and value.}

\item{collapse}{a character used to collapse definitions into a
single string.}
}
\description{
Return formatted notes from a tex glossary file or glossary object
}
\examples{
file <- system.file("glo", "glossary.tex", package = "pmtables")

glossary_notes(file, WT, CRCL)

g <- as_glossary(ss = "steady state", ALB = "albumin", WT = "weight")

glossary_notes(g, ALB, ss)

what <- c("WT", "NCA", "NPDE")

g <- read_glossary(file)

glossary_notes(g, tidyselect::all_of(what))

l <- list(ss = "steady state", ALB = "albumin", WT = "weight")

glossary_notes(l, ALB, ss)

}
\seealso{
\code{\link[=st_notes_glo]{st_notes_glo()}}, \code{\link[=read_glossary]{read_glossary()}}
}
