% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-utils.R
\name{grep_df}
\alias{grep_df}
\alias{grepl_df}
\title{Find matching rows in a data frame}
\usage{
grep_df(data, pattern, cols = names(data))

grepl_df(data, pattern, cols = names(data))
}
\arguments{
\item{data}{a data frame to search}

\item{pattern}{a regular expression}

\item{cols}{a character vector of column names to search}
}
\description{
Find matching rows in a data frame
}
