% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-tabular.R
\name{make_tabular}
\alias{make_tabular}
\title{Create tabular environment from data frame}
\usage{
make_tabular(data, prime_fun = tab_prime, escape_fun = tab_escape, ...)
}
\arguments{
\item{data}{a data.frame}

\item{prime_fun}{a function to prime the data frame for rendering in TeX}

\item{escape_fun}{a function to escape characters that have special meaning
in TeX}

\item{...}{not used}
}
\description{
Create tabular environment from data frame
}
