% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-digits.R
\name{new_digits}
\alias{new_digits}
\title{Create a digits object}
\usage{
new_digits(.fun = sig, .default = 3, ..., .data = NULL)
}
\arguments{
\item{.fun}{function for modifying digits on a number}

\item{.default}{the default value for digits}

\item{...}{\code{name=value} pairs, where \code{name} references the data column name
and \code{value} is the number of digits for that columns}

\item{.data}{a named list to be used in place of \code{...}}
}
\description{
The idea here is to capture three pieces of data: (1) a function
for formatting digits in an output number (2) a default number of
digits to use and (3) a named list with digit customization
information; the names correspond to potential future column
names and the values are custom digit information.
}
\examples{

x <- new_digits(round, .default = 2, WT = 1, ALB = 3)

as.list(x)

}
