% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{pmt.obs}
\alias{pmt.obs}
\title{analysis1 data set - observations only}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3530 rows and 36 columns.
}
\usage{
pmt.obs
}
\description{
analysis1 data set - observations only
}
\keyword{datasets}
