% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{pmt.pk}
\alias{pmt.pk}
\title{analysis1 data set - PK observations only}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3210 rows and 36 columns.
}
\usage{
pmt.pk
}
\description{
analysis1 data set - PK observations only
}
\keyword{datasets}
