% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{pmt_first}
\alias{pmt_first}
\title{analysis1 data set - first record only}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 160 rows and 36 columns.
}
\usage{
pmt_first
}
\description{
analysis1 data set - first record only
}
\keyword{datasets}
