% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{pmt_obs}
\alias{pmt_obs}
\title{analysis1 data set - observations only}
\format{
A data frame with 3530 rows and 36 variables
}
\usage{
pmt_obs
}
\description{
These are observation records (\code{EVID=0}) from the
\link{analysis1} data set.
}
\seealso{
\link{analysis1}, \link{pmt_first}, \link{pmt_pk}
}
\keyword{datasets}
