% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{pmt_pk}
\alias{pmt_pk}
\title{analysis1 data set - PK observations only}
\format{
A data frame with 3210 rows and 36 variables
}
\usage{
pmt_pk
}
\description{
These are PK observation records from the \link{analysis1}
data set.
}
\seealso{
\link{analysis1}, \link{pmt_first}, \link{pmt_obs}
}
\keyword{datasets}
