% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{pmt_summarized}
\alias{pmt_summarized}
\title{analysis1 data set - summarized}
\format{
A data frame with 13 rows and 9 variables
}
\usage{
pmt_summarized
}
\description{
This is an example of a summary of the \link{analysis1}
data set.
}
\keyword{datasets}
