% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{print.pt_opts}
\alias{print.pt_opts}
\title{Print the pt_opts object}
\usage{
\method{print}{pt_opts}(x, ...)
}
\arguments{
\item{x}{the \code{pt_opts} object}

\item{...}{not used}
}
\description{
Print the pt_opts object
}
\examples{
pt_opts

}
