% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{print.stobject}
\alias{print.stobject}
\alias{names.stobject}
\title{Methods for stobject}
\usage{
\method{print}{stobject}(x, ...)

\method{names}{stobject}(x)
}
\arguments{
\item{x}{an stobject}

\item{...}{other arguments}
}
\description{
Methods for stobject
}
