% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-cat-cat.R
\name{pt_cat_cat}
\alias{pt_cat_cat}
\alias{pt_cat_cat_style}
\alias{pt_cat_cat_data}
\title{Categorical by categorical table}
\usage{
pt_cat_cat(data, cols_down, cols_across)

pt_cat_cat_style(tab)

pt_cat_cat_data(data, cols_down, cols_across)
}
\description{
Categorical by categorical table
}
