% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_long}
\alias{pt_cat_long}
\title{Discrete data summary in long format}
\usage{
pt_cat_long(
  data,
  cols,
  span = ".total",
  all_name = " ",
  all_name_span = "Summary",
  summarize = c("both", "right", "top", "none"),
  table = NULL,
  by = NULL,
  denom = c("group", "total")
)
}
\arguments{
\item{data}{the data frame to summarize; the user should filter or subset
so that \code{data} contains exactly the records to be summarized; pmtables will
not add or remove rows prior to summarizing \code{data}}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{span}{variable name for column spanner}

\item{all_name}{a name to use for the complete data summary}

\item{all_name_span}{table column name to use for data summaries across
levels of \code{span} if it is provided}

\item{summarize}{where to include a data summary across subgroups;
use \code{none} to drop the summary from the table}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{by}{use \code{span} argument instead}

\item{denom}{the denominator to use when calculating percent for each level;
\code{group} uses the total number in the chunk being summarized; \code{total} uses
the total number in the data set; historically, \code{group} has been used as the
default.}
}
\value{
An object with class \code{pmtable}; see \link{class-pmtable}.
}
\description{
Discrete data summary in long format
}
\details{
The data summary for all cells in the table is \code{count (percent)}. The number
of data records in each column variable level is given under the column
title as \code{n}.

When \code{group} is selected for \code{denom}, \code{percent} is calculated with
denominator set to \code{n}, the total for each column variable level. When
\code{total} is selected for \code{denom}, then \code{percent} is calculated by the total
number of records in the input data.

The notes in this table are generated with \code{\link[=pt_cat_long_notes]{pt_cat_long_notes()}}.
}
\examples{

out <- pt_cat_long(pmt_first, cols = "SEXf,ASIANf", span = "FORMf")

\dontrun{
st2report(stable(out))
}

}
