% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_long_notes}
\alias{pt_cat_long_notes}
\title{Return table notes for pt_cat_long}
\usage{
pt_cat_long_notes(include_n = TRUE, note_add = NULL)
}
\arguments{
\item{include_n}{if \code{TRUE}, add a note for \code{n}}

\item{note_add}{additional notes to include}
}
\description{
See \code{\link[=pt_cat_long]{pt_cat_long()}}.
}
