% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_study}
\alias{pt_cat_study}
\title{Discrete covariate table by study}
\usage{
pt_cat_study(
  data,
  cols,
  study_col = vars(`Study ID` = STUDY),
  summarize_all = TRUE,
  all_name = "All studies",
  table = NULL,
  note = NULL,
  wide = FALSE
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{study_col}{character name of the study ID column; passed to
\link{pt_cont_long} as \code{panel} or \link{pt_cont_wide} as \code{by}}

\item{wide}{if \code{TRUE} covariates are rendered west to east; if \code{FALSE} then
they are rendered north to south}

\item{...}{other arguments passed to \link{pt_cat_long} or \link{pt_cat_wide}}
}
\description{
Discrete covariate table by study
}
\examples{

data <- pmtables:::data("id")

pt_cat_study(data,cols = "SEXf,FORMf", study="STUDYf")

pt_cat_study(data,cols = "SEXf,FORMf", study="STUDYf", wide = TRUE)

}
