% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_wide}
\alias{pt_cat_wide}
\title{Discrete data summary in long format}
\usage{
pt_cat_wide(
  data,
  cols,
  by = ".total",
  panel = by,
  table = NULL,
  all_name = "All data",
  summarize = c("bottom", "none"),
  complete = FALSE
)
}
\arguments{
\item{data}{the data frame to summarize; the user should filter or subset
so that \code{data} contains exactly the records to be summarized; pmtables will
not add or remove rows prior to summarizing \code{data}}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{by}{a grouping variable for the summary; may be given as character
vector or quosure.}

\item{panel}{data set column name to stratify the summary}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{all_name}{a name to use for the complete data summary}

\item{summarize}{where to put an all-data summary; choose \code{none} to omit the
summary from the table.}

\item{complete}{logial; if \code{TRUE}, then data the summary will be completed
for missing levels of \code{by}and \code{panel}.}
}
\value{
An object with class \code{pmtable}; see \link{class-pmtable}.
}
\description{
Discrete data summary in long format
}
\details{
The data summary for this table is \code{count (percent)}. The number of
data points for each row is also summarized as \code{n} on the left hand side
of the table (either on the far left or just to the right of the \code{by}
column).

The notes in this table are generated with \code{\link[=pt_cat_wide_notes]{pt_cat_wide_notes()}}.
}
\examples{
out1 <- pt_cat_wide(pmt_first, cols = "SEXf,ASIANf")
stable(out1)

out2 <- pt_cat_wide(pmt_first, cols = "SEXf,ASIANf", by = "FORMf")
stable(out2)

\dontrun{
st2report(stable(out1))
st2report(stable(out2))
}

}
