% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_wide_notes}
\alias{pt_cat_wide_notes}
\title{Return table notes for pt_cat_wide}
\usage{
pt_cat_wide_notes(include_n = TRUE, note_add = NULL)
}
\arguments{
\item{include_n}{include the note for \code{n} column}

\item{note_add}{additional notes to include}
}
\description{
See \code{\link[=pt_cat_wide]{pt_cat_wide()}}.
}
