% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{pt_cont_long}
\alias{pt_cont_long}
\title{Continuous data summary in long format}
\usage{
pt_cont_long(
  data,
  cols,
  panel = ".total",
  table = NULL,
  units = NULL,
  note = NULL,
  summarize_total = TRUE,
  all_name = "All data",
  fun = df_sum_2
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{columns to summarize; may be character vector or quosure}

\item{panel}{data set column name to stratify the summary}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{units}{a named list to use for unit lookup (see details and examples)}

\item{note}{text to append to the bottom of the table using
\link[mrggt:tab_source_note]{mrggt::tab_source_note}}

\item{summarize_total}{if \code{TRUE} then a complete data summary will be appended
to the bottom of the table}

\item{all_name}{a name to use for the complete data summary}
}
\description{
Continuous data summary in long format
}
\examples{
data <- pmtables:::data("id")

ans <- pt_cont_long(data, cols = .cols(WT,ALB,CRCL))

ans <- pt_cont_long(data, cols = "WT,CRCL", panel = "SEXf")

}
