% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{pt_cont_long_notes}
\alias{pt_cont_long_notes}
\title{Return table notes for pt_cont_long}
\usage{
pt_cont_long_notes(note_add = NULL)
}
\arguments{
\item{note_add}{additional notes to include}
}
\description{
See \code{\link[=pt_cont_long]{pt_cont_long()}}.
}
