% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{pt_cont_study}
\alias{pt_cont_study}
\title{Continuous covariate table by study}
\usage{
pt_cont_study(data, cols, study_col = "STUDY", wide = FALSE, ...)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{study_col}{character name of the study ID column}

\item{wide}{if \code{TRUE} covariates are rendered west to east; if \code{FALSE} then
they are rendered north to south}

\item{...}{other arguments passed to \link{pt_cont_long} or \link{pt_cont_wide}}
}
\description{
Continuous covariate table by study
}
\examples{

data <- pmtables:::data("id")

pt_cont_study(data,cols = "WT,ALB,SCR", study="STUDYf")

pt_cont_study(data,cols = "WT,ALB,SCR", study="STUDYf", wide = TRUE)

}
