% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{pt_cont_wide}
\alias{pt_cont_wide}
\title{Create a continuous data summary table in wide format}
\usage{
pt_cont_wide(
  data,
  cols,
  by = ".total",
  panel = by,
  table = NULL,
  units = NULL,
  digits = NULL,
  all_name = "All data",
  fun = str_sum_2,
  panel.label.add = pt_opts$panel.label.add
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{by}{a grouping variable name}

\item{panel}{a variable for paneling the summary}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{units}{a named list to use for unit lookup (see details and examples)}

\item{digits}{a \code{digits} object (see \link{new_digits})}

\item{all_name}{a name to use for the complete data summary}

\item{fun}{the data summary function (see details)}

\item{panel.label.add}{logical; if \code{TRUE}, panel variable name will be pasted
to the panel variable value}
}
\description{
Create a continuous data summary table in wide format
}
