% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{pt_cont_wide_notes}
\alias{pt_cont_wide_notes}
\title{Return table notes for pt_cont_wide}
\usage{
pt_cont_wide_notes(note_add = NULL)
}
\arguments{
\item{note_add}{additional notes to include}
}
\description{
See \code{\link[=pt_cont_wide]{pt_cont_wide()}}.
}
