% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{pt_data_inventory}
\alias{pt_data_inventory}
\title{Create a data inventory table}
\usage{
pt_data_inventory(
  data,
  outer = ".total",
  inner = outer,
  inner_summary = TRUE,
  drop_miss = FALSE,
  stacked = FALSE,
  table = NULL,
  all_name = "all",
  ...
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{outer}{the outer grouping variable; may be character or quosure}

\item{inner}{the inner grouping variable; may be character or quosure}

\item{inner_summary}{if \code{TRUE}, then a summary of the inner variable will
be provided}

\item{drop_miss}{if \code{TRUE}, then \code{MISS} will be dropped, but only when all
\code{MISS} values are equal to zero}

\item{stacked}{if \code{TRUE}, then independent summaries are created by \code{outer}
and included in a single table (see examples)}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{all_name}{a name to use for the complete data summary}

\item{...}{other arguments passed to \link{pt_data_inventory}}
}
\description{
Output columns include counts for subjects, observations, BQL observations,
and missing observations and percentage of observations that are BQL.
}
\examples{
data <- pmtables:::data("id")

ans <- pt_data_inventory(data, outer = .cols("Renal function" = RFf))


}
