% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{pt_data_inventory_cat}
\alias{pt_data_inventory_cat}
\title{Data inventory by covariate}
\usage{
pt_data_inventory_cat(
  data,
  cols,
  drop_miss = FALSE,
  table = NULL,
  summarise_all = TRUE,
  all_name = "All data",
  dv_col = "DV",
  bq_col = find_bq_col(data),
  id_col = "ID"
)
}
\description{
Data inventory by covariate
}
