% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{pt_data_study}
\alias{pt_data_study}
\title{Data inventory by study}
\usage{
pt_data_study(data, study_col = "STUDY", panel = study_col, ...)
}
\arguments{
\item{data}{the data frame to summarize}

\item{study_col}{the name of the column containing the study identifier;
may be characer or quosure (see \link[dplyr:vars]{dplyr::vars})}

\item{panel}{another categorical data set column name to stratify the
data summary}

\item{...}{other arguments passed to \link{pt_data_inventory} and
\link{data_inventory_chunk}}
}
\description{
This is a convenience wrapper around \link{pt_data_inventory}
}
\examples{
data <- pmtables:::data("id")

ans <- pt_data_study(data, study_col = "STUDYf")

}
