% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics-table.R
\name{pt_demographics}
\alias{pt_demographics}
\title{This is the title}
\usage{
pt_demographics(
  data,
  cols_cont,
  cols_cat,
  span,
  stat_name = "Statistic",
  stat_width = 2
)
}
\arguments{
\item{cols_cat}{}
}
\value{
An object of class \code{pmtable}
}
\description{
This function makes a single table from both continuous and categorical data.
}
\examples{

}
\seealso{
\code{\link[=pt_cont_long]{pt_cont_long()}}, \code{\link[=pt_cat_long]{pt_cat_long()}}
}
