% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-functions.R
\name{pt_wrap}
\alias{pt_wrap}
\title{Wrap stable output in table environment}
\usage{
pt_wrap(x, center = TRUE, caption = NULL, con = stdout())
}
\arguments{
\item{x}{tabular text}

\item{center}{if \code{TRUE}, center the table}

\item{caption}{the table caption}

\item{con}{where to write the output}
}
\description{
Wrap stable output in table environment
}
