% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_glossary}
\alias{read_glossary}
\title{Read and parse a tex glossary file}
\usage{
read_glossary(file)
}
\arguments{
\item{file}{path to the glossary file}
}
\value{
A named list.
}
\description{
Read and parse a tex glossary file
}
\examples{
file <- system.file("tex", "glossary.tex", package = "pmtables")

x <- read_glossary(file)

names(x)

x$WT

x$SC

}
