% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{select_glossary}
\alias{select_glossary}
\title{Select entries from a glossary object}
\usage{
select_glossary(x, ...)
}
\arguments{
\item{x}{a glossary object.}

\item{...}{tidy-select syntax naming glossary entry labels to select.}
}
\value{
A new glossary object with selected entries.
}
\description{
Select entries from a glossary object
}
\examples{
gl <- as_glossary(list(a = "apple", b = "banana", c = "cat", d = "dog"))

select_glossary(gl, b, d)

}
