% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_as_image}
\alias{st_as_image}
\title{Display table from pdf image}
\usage{
st_as_image(x, width = getOption("pmtables.image.width", 0.8), ...)
}
\arguments{
\item{x}{an stable object.}

\item{width}{the relative width of the image; passed to \code{\link[=st_image_show]{st_image_show()}};
this must be greater than 0 and less than or equal to 1.}

\item{...}{arguments passed to \code{\link[=st_aspdf]{st_aspdf()}}.}
}
\value{
A possibly resized \code{magick} image object (see \code{\link[magick:editing]{magick::image_read_pdf()}}).
}
\description{
Use this function to run an stable object in to a high-quality pdf file
and then include it into a html document.
}
\details{
This function depends on the magick and pdftools packages being installed.
\enumerate{
\item First, \code{x} is rendered with \code{\link[=st_aspdf]{st_aspdf()}}
\item Next, the pdf file is read using \code{\link[magick:editing]{magick::image_read_pdf()}}
\item Finally, the image is possibly resized via \code{\link[=st_image_show]{st_image_show()}}
}
}
