% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_blank}
\alias{st_blank}
\title{Add column blank information to st object}
\usage{
st_blank(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{quoted or unquoted column names to be passed to \code{\link[=stable]{stable()}} as
\code{cols_blank}}
}
\description{
See the \code{cols_blank} argument passed to \code{\link[=stable]{stable()}} and then \code{\link[=tab_cols]{tab_cols()}}.
This function can be called multiple times and will accumulate \code{cols_blank}
data.
}
