% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_bold}
\alias{st_bold}
\alias{st_it}
\title{Make columns bold or italics}
\usage{
st_bold(x, cols, pattern = "*")

st_it(x, cols, pattern = "*")
}
\arguments{
\item{x}{an stobject}

\item{cols}{columns to make bold}

\item{pattern}{passed to \code{\link[=tex_bold]{tex_bold()}} or \code{\link[=tex_it]{tex_it()}}}
}
\description{
Make columns bold or italics
}
