% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_caption}
\alias{st_caption}
\title{Set table caption}
\usage{
st_caption(x, text, ...)
}
\arguments{
\item{x}{an stobject.}

\item{text}{caption text to be parsed.}

\item{...}{additional arguments passed to \code{\link[=as.caption]{as.caption()}}; see details.}
}
\description{
Set table caption
}
\details{
A short title for the table can be specified in one of two ways. First, the
short title can be specified by including it in brackets (\verb{[]}) as the
first characters in \code{text}. Alternatively, short title text can get passed
separately into \code{st_caption()} under the argument \code{short}, which is
is eventually handled by \code{\link[=as.caption]{as.caption()}}.

By default, \code{\link[=as.caption]{as.caption()}} will repeat the short title at the start of the
main caption text. The user can pass \code{repeat_short = FALSE} into
\code{st_caption()} (and eventually to \code{\link[=as.caption]{as.caption()}}) to suppress this behavior.
}
\examples{
tab <- st_new(stdata())
tab <- st_caption(tab, "[Full covariate model estimates]. Run number 101.")
tab$caption

text <- stable(tab)
text

}
