% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_clear_reps}
\alias{st_clear_reps}
\alias{st_clear_grouped}
\title{Add clear rep information to st object}
\usage{
st_clear_reps(x, ..., .now = FALSE)

st_clear_grouped(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{quoted or unquoted column names passed to \code{\link[=stable]{stable()}} as
\code{clear_reps}}

\item{.now}{if \code{TRUE}, the data is immediately processed; otherwise, the
processing is done after the pipeline completes}
}
\description{
See the \code{clear_reps} argument passed to \code{\link[=stable]{stable()}} and then to
\code{\link[=tab_clear_reps]{tab_clear_reps()}}.
}
