% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_clone}
\alias{st_clone}
\title{Clone an stobject}
\usage{
st_clone(x)
}
\arguments{
\item{x}{an stobject object.}
}
\value{
A copy (\code{y}) of \code{x} such that \code{y} can be modified without modifying \code{x}.
}
\description{
Clone an stobject
}
\examples{
x <- st_new(stdata())
y <- st_clone(x)

y$data$STUDY <- NULL
x$data
y$data

# Get back to a regular environment
class(x) <- "environment"
class(y) <- "environment"

x
y

}
