% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-control.R
\name{st_control}
\alias{st_control}
\alias{st_control.default}
\alias{st_control.stobject}
\title{Table control parameters}
\usage{
st_control(...)

\method{st_control}{default}(
  row_space = 1.4,
  col_space = 5,
  lt_row_space = 0.3,
  fontsize = "normalsize",
  header_row = 0.56,
  note_type = c("tpt", "minipage"),
  note_hline = c("top", "bottom", "both", "none"),
  note_hline_pt = 0.4,
  note_table_skip = 0.67,
  note_skip = 0.02,
  note_sanitize = TRUE,
  note_escape = "_",
  ...
)

\method{st_control}{stobject}(x, ...)
}
\arguments{
\item{...}{used only in the generic}

\item{row_space}{multiplier to increase or decrease padding between
table rows; for example, \code{row_space = 1.4} means increase by 40\%}

\item{col_space}{padding between columns; unit is \code{pt}}

\item{lt_row_space}{\strong{extra} space between rows of longtable;
unit is \code{em}}

\item{fontsize}{font size}

\item{header_row}{amount of space that will be \strong{removed} between
rows of the table header; unit is \code{em}; this must be a positive
number}

\item{note_type}{\code{tpt} puts notes in the third part of \code{threeparttable} and
\code{minipage} puts notes in a minipage below the table}

\item{note_hline}{where to place hlines when \code{minipage} notes are used}

\item{note_hline_pt}{pointsize for \code{minipage} note \code{hline}}

\item{note_table_skip}{vertical space between the bottom of the table
and the start of \code{minipage} notes; unit is \code{cm}}

\item{note_skip}{vertical space after the top \code{hline} where the \code{minipage}
notes start}

\item{note_sanitize}{if \code{TRUE}, notes will be sanitized}

\item{note_escape}{a vector of characters to sanitize}

\item{x}{an stobject}
}
\description{
Table control parameters
}
