% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_drop}
\alias{st_drop}
\title{Drop data columns}
\usage{
st_drop(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{column names to drop}
}
\description{
See the \code{drop} argument to \code{\link[=stable]{stable()}}.
}
