% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_edit}
\alias{st_edit}
\alias{tab_edit}
\title{Edit table contents}
\usage{
st_edit(x, ...)

tab_edit(data, pattern, replacement, cols = names(data))
}
\arguments{
\item{x}{an stobject}

\item{pattern}{passed to \code{\link[stringr:str_replace]{stringr::str_replace()}}}

\item{replacement}{passed to \code{\link[stringr:str_replace]{stringr::str_replace()}}}

\item{cols}{data columns to check for \code{pattern}}
}
\description{
Edit table contents
}
