% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_files}
\alias{st_files}
\title{Add file name information to st object}
\usage{
st_files(
  x,
  r = getOption("mrg.script", NULL),
  output = NULL,
  output_dir = getOption("pmtables.dir"),
  esc = NULL
)
}
\arguments{
\item{x}{an stobject.}

\item{r}{set \code{r_file}, passed to \code{\link[=stable]{stable()}}.}

\item{output}{set \code{output_file}, the \code{.tex} file containing rendered
table code; this may be the base file name or the full path to the
file; passed to \code{\link[=stable]{stable()}}.}

\item{output_dir}{directory for \code{output_file}; this will be overridden with
a warning if \code{output_file} contains the full path to the file.}

\item{esc}{passed to \code{\link[=tab_escape]{tab_escape()}}; use \code{NULL} to bypass escaping.}
}
\description{
See the \code{r_file} and \code{output_file} arguments passed to \code{\link[=stable]{stable()}} and then
to \code{\link[=tab_notes]{tab_notes()}}.
}
\examples{
library(dplyr)

ob <- st_new(ptdata())

ob \%>\% st_files(r = "foo.R", output = "foo.tex") \%>\% stable()

}
