% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_hline}
\alias{st_hline}
\title{Add hline information to st object}
\usage{
st_hline(x, at = NULL, from = NULL)
}
\arguments{
\item{x}{and stobject}

\item{at}{logical or integer locations for hline passed to \code{\link[=stable]{stable()}} as
\code{hline_at}}

\item{from}{character column name used to divide a table; passed to
\code{\link[=stable]{stable()}} as \code{hline_from}}
}
\description{
See the \code{hline_at} and \code{hline_from} arguments to \code{\link[=stable]{stable()}}.
}
