% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_image_show}
\alias{st_image_show}
\title{Read table output that has been saved to pdf or png format}
\usage{
st_image_show(path, width = 0.95, knitting = getOption("knitr.in.progress"))
}
\arguments{
\item{path}{path to a png or pdf file saved on disk.}

\item{width}{for resizing the magick object; depends on context.}

\item{knitting}{if \code{TRUE}, the context is assummed to be document knit (see
default) and the magick object is returned after resizing; if \code{FALSE}, then
interactive context is assummed and the magick object is returned after
resizing according to the current device width.}
}
\description{
This function requires the magick and pdftools packages to be installed.
}
