% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_inline_pdf}
\alias{st_inline_pdf}
\title{Display table from pdf image}
\usage{
st_inline_pdf(x, width = getOption("pmtables.st_inline.width", 0.8), ...)
}
\arguments{
\item{x}{an stable object.}

\item{width}{the width as percent; passed to
\code{\link[magick:geometry]{magick::geometry_size_percent()}}.}

\item{...}{arguments passed to \code{\link[=st_aspdf]{st_aspdf()}}.}
}
\value{
A possibly resized image object (see \code{\link[magick:editing]{magick::image_read_pdf()}}).
}
\description{
Use this function to run an stable object in to a high-quality pdf file
and then include it into a html document.
}
\details{
This function depends on the magick and pdftools packages being installed.
\enumerate{
\item First, \code{x} is rendered with \code{\link[=st_aspdf]{st_aspdf()}}
\item Next, the pdf file is read using \code{\link[magick:editing]{magick::image_read_pdf()}}
\item Finally, the image is resized if \code{width} is passed.
}
}
