% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_make}
\alias{st_make}
\title{Convert st object to table output}
\usage{
st_make(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{other arguments passed to \code{\link[=stable]{stable()}}}
}
\description{
Convert st object to table output
}
