% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_new}
\alias{st_new}
\title{Create an st object}
\usage{
st_new(data)
}
\arguments{
\item{data}{the data frame to pass to \code{\link[=stable]{stable()}}}
}
\description{
The st object will collect various configuration settings and pass those
to \code{\link[=stable]{stable()}} when the object is passed to \code{\link[=st_make]{st_make()}}.
}
