% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_noteconf}
\alias{st_noteconf}
\title{Add note config information to st object}
\usage{
st_noteconf(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{named arguments passed to \code{\link[=noteconf]{noteconf()}}}
}
\description{
See the \code{note_config} argument to \code{\link[=stable]{stable()}}.
}
