% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_notes}
\alias{st_notes}
\alias{st_notes_ap}
\title{Add note information to st object}
\usage{
st_notes(
  x,
  ...,
  esc = NULL,
  config = NULL,
  collapse = "; ",
  append = FALSE,
  replace = FALSE,
  to_string = FALSE
)

st_notes_ap(...)
}
\arguments{
\item{x}{an stobject}

\item{...}{table notes}

\item{esc}{passed to \code{\link[=tab_escape]{tab_escape()}}; use \code{NULL} to bypass escaping the notes}

\item{config}{named list of arguments for \code{\link[=noteconf]{noteconf()}}}

\item{collapse}{a character string to separate notes which are pasted
together when flattening or appending; this should usually end in a single
space (see default).}

\item{append}{logical; if \code{TRUE}, then incoming notes are appended to the
previous, single note in the notes list. When \code{...} contains multiple
notes, then the notes are pasted together first.}

\item{replace}{logical; if \code{TRUE}, remove any existing notes before procesing
new notes.}

\item{to_string}{logical; if \code{TRUE}, then all notes are collapsed to a single
string.}
}
\description{
See the \code{notes} and \code{note_config} arguments passed to \code{\link[=stable]{stable()}} and then to
\code{\link[=tab_notes]{tab_notes()}}. The function can be called multiple times and can accumulate
\code{notes} data in various ways.
}
\examples{
library(dplyr)

ob <- st_new(ptdata())

ob \%>\% st_notes("ALB: albumin (g/dL)") \%>\% stable()

}
