% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_notes_app}
\alias{st_notes_app}
\title{Append a note to the previous position of a note vector}
\usage{
st_notes_app(...)
}
\arguments{
\item{...}{passed to \code{\link[=st_notes]{st_notes()}}.}
}
\description{
Append a note to the previous position of a note vector
}
\details{
Note that the call to \code{\link[=st_notes]{st_notes()}} will force in the argument
\code{append = TRUE}.
}
