% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_notes_detach}
\alias{st_notes_detach}
\title{Detach table notes from the table}
\usage{
st_notes_detach(x, width = 0.8, type = "minipage", ...)
}
\arguments{
\item{x}{an stobject.}

\item{width}{passed to \code{\link[=noteconf]{noteconf()}} via \code{\link[=st_noteconf]{st_noteconf()}}.}

\item{type}{passed to \code{\link[=noteconf]{noteconf()}} via \code{\link[=st_noteconf]{st_noteconf()}}; this argument should
not be changed if detached notes are desired.}

\item{...}{other arguments passed to \code{\link[=noteconf]{noteconf()}} via \code{\link[=st_noteconf]{st_noteconf()}}.}
}
\value{
An updated object with class \code{stobject}, which can be piped to other
functions.
}
\description{
Detached notes are rendered underneath the table, in a separate minipage
format. By default, there is an \code{hline} renered between the table and the
notes. It is common to adjust the width of the minipage holding the notes
depending on the width of the table and the extent of the notes.
}
