% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_notes_glo}
\alias{st_notes_glo}
\title{Add table notes based on acronyms from a tex glossary file}
\usage{
st_notes_glo(
  x,
  glossary,
  ...,
  sep = ": ",
  collapse = "; ",
  labels = NULL,
  width = NULL
)
}
\arguments{
\item{x}{an stobject.}

\item{glossary}{a named list generated from \code{\link[=read_glossary]{read_glossary()}}.}

\item{...}{unquoted names matching those names in \code{glossary}.}

\item{sep}{character to separate name and value.}

\item{collapse}{a character used to collapse definitions into a
single string.}

\item{labels}{a character vector or comma-separates string of definition
labels to get appended to the end of names passed in \code{...}.}

\item{width}{if numeric, \code{\link[=st_notes_detach]{st_notes_detach()}} will be called with \code{width}
argument.}
}
\description{
Add table notes based on acronyms from a tex glossary file
}
\examples{
library(dplyr)

file <- system.file("glo", "glossary.tex", package = "pmtables")

x <- read_glossary(file)

st_new(stdata()) \%>\%
  st_notes_glo(x, WT, CRCL, SCR, width = 1) \%>\%
  stable()

}
\seealso{
\code{\link[=glossary_notes]{glossary_notes()}}, \code{\link[=read_glossary]{read_glossary()}}
}
