% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_notes_rm}
\alias{st_notes_rm}
\title{Remove notes from the table}
\usage{
st_notes_rm(x)
}
\arguments{
\item{x}{an st object.}
}
\value{
An updated object with class \code{stobject}, which can be piped to other
functions.
}
\description{
The can be useful when manipulating an object from one of the pmtable
functions (e.g. \code{\link[=pt_cont_long]{pt_cont_long()}} or \code{\link[=pt_demographics]{pt_demographics()}}, when notes are
automatically added to the table.
}
