% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_notes_str}
\alias{st_notes_str}
\title{Convert existing note vector into a single string}
\usage{
st_notes_str(x, collapse = "; ")
}
\arguments{
\item{x}{an stobject.}

\item{collapse}{a character string to separate notes which are pasted
together when flattening or appending; this should usually end in a single
space (see default).}
}
\value{
An updated object with class \code{stobject}, which can be piped to other
functions.
}
\description{
Convert existing note vector into a single string
}
