% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_notes_sub}
\alias{st_notes_sub}
\title{Edit lines in table notes}
\usage{
st_notes_sub(x, where, replacement, fixed = FALSE)
}
\arguments{
\item{x}{an stobject.}

\item{where}{a regular expression for finding a line in table notes to
replace; alternatively, this can be an integer specifying the line to
replace.}

\item{replacement}{the replacement text for line matching by \code{where}.}

\item{fixed}{passed to \code{\link[=grep]{grep()}} when \code{where} is character.}
}
\value{
An updated object with class \code{stobject}, which can be piped to other
functions.
}
\description{
This function allows the replacement of \emph{an entire line} in table notes.
The line which is replaced is matched by a regular expression or identified
directly with the integer position  in the notes vector to replace.
}
\details{
A warning is generated if there are no notes already existing in \code{x}. A
warning is also generated if a regular expression fails to match any lines.
In case multiple lines are matched, only the first matching line is
substituted.
}
