% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_panel}
\alias{st_panel}
\title{Add panel information to st object}
\usage{
st_panel(x, ...)
}
\arguments{
\item{...}{passed to \code{\link[=rowpanel]{rowpanel()}}}

\item{an}{stobject}
}
\description{
Add panel information to st object
}
