% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{st_preview}
\alias{st_preview}
\title{Preview an stable object}
\usage{
st_preview(x, ...)
}
\arguments{
\item{x}{an stable object}

\item{...}{passed to \code{\link[texPreview:tex_preview]{texPreview::tex_preview()}}}
}
\description{
Preview an stable object
}
