% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_rename}
\alias{st_rename}
\title{Add column rename information to st object}
\usage{
st_rename(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{column rename items in \code{new-name = old-name} format; passed
to \code{\link[=stable]{stable()}} as \code{col_rename}}
}
\description{
See the \code{col_rename} argument to \code{\link[=stable]{stable()}}. This function can be called
multiple times and will accumulate \code{col_rename} data.
}
