% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_sizes}
\alias{st_sizes}
\title{Add table size information to st object}
\usage{
st_sizes(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{passed to \code{\link[=tab_size]{tab_size()}}}
}
\description{
See the \code{sizes} argument to \code{\link[=stable]{stable()}}, passed to \code{\link[=tab_size]{tab_size()}}.
}
