% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_space}
\alias{st_space}
\title{Add row and column spacing information to st object}
\usage{
st_space(x, row = NULL, col = NULL)
}
\arguments{
\item{row}{set \code{row_space}, passed to \code{\link[=stable]{stable()}}}

\item{col}{set \code{col_space}, passed to \code{\link[=stable]{stable()}}}
}
\description{
Add row and column spacing information to st object
}
