% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_span_split}
\alias{st_span_split}
\title{Add column split spanner information to st object}
\usage{
st_span_split(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{passed to \code{\link[=colsplit]{colsplit()}}}
}
\description{
Add column split spanner information to st object
}
