% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_units}
\alias{st_units}
\title{Add unit information to st object}
\usage{
st_units(x, ..., parens = TRUE)
}
\arguments{
\item{x}{an stobject}

\item{...}{named items of the form \code{COL = unit} or a named list
of units}

\item{parens}{if \code{TRUE}, parens will be added to any unit whose first
character is not \code{(}}
}
\description{
See the \code{units} argument to \code{\link[=stable]{stable()}}. Units can be passed either as
\code{name=value} pairs or as a named list with \code{\link[=st_args]{st_args()}}. Units can
alternatively be passed as an argument to \code{\link[=stable]{stable()}} as a pre-formed, named
list using \code{\link[=st_args]{st_args()}}.  Passing as an argument this way will overwrite units
specified with \code{\link[=st_units]{st_units()}}. It is recommended to use either \code{\link[=st_units]{st_units()}} or
\code{\link[=st_args]{st_args()}} but not both.
}
