% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-dependencies.R
\name{st_use_knit_deps}
\alias{st_use_knit_deps}
\alias{st_knit_deps}
\alias{st_using_knit_deps}
\title{Invoke latex dependencies}
\usage{
st_use_knit_deps(packages = NULL, force = FALSE)

st_knit_deps()

st_using_knit_deps()
}
\arguments{
\item{packages}{a character vector of latex package names; will get passed
to \code{\link[rmarkdown:latex_dependency]{rmarkdown::latex_dependency()}} and then to \code{\link[knitr:knit_meta]{knitr::knit_meta_add()}};
default is to install packages returned by \code{\link[=st_knit_deps]{st_knit_deps()}}}

\item{force}{if \code{TRUE}, then add latex dependencies regardless of
knit status; this is for internal testing use only}
}
\value{
Returns invisible \code{NULL}.
}
\description{
Call this function when knitting a document that includes pmtables output.
}
\details{
Default set of packages that will be invoked \strong{when knitting a document}:
\itemize{
\item \code{float}
\item \code{booktabs}
\item \code{longtable}
\item \code{threeparttable}
\item \code{pdflscape}
\item \code{array}
\item \code{caption}
}
}
