% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-dependencies.R
\name{st_usepackage}
\alias{st_usepackage}
\alias{st_packages}
\alias{st_using_deps}
\title{Invoke latex dependencies}
\usage{
st_usepackage(packages = NULL)

st_packages()

st_using_deps()
}
\arguments{
\item{packages}{a character vector of latex package names; will get passed
to \code{\link[rmarkdown:latex_dependency]{rmarkdown::latex_dependency()}} and then to \code{\link[knitr:knit_meta]{knitr::knit_meta_add()}}.
The default is to install packates returned by \code{\link[=st_packages]{st_packages()}}.}
}
\value{
Returns invisible \code{NULL}.
}
\description{
Call this function when knitting a document that includes pmtables output.
}
