% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-long.R
\name{stable_long}
\alias{stable_long}
\alias{stable_long.data.frame}
\alias{stable_long.stobject}
\alias{stable_long.pmtable}
\title{Create longtable output from an R data frame}
\usage{
stable_long(data, ...)

\method{stable_long}{data.frame}(
  data,
  note_config = noteconf(type = "minipage"),
  inspect = FALSE,
  lt_cap_macro = "",
  lt_cap_text = "",
  lt_cap_short = "",
  lt_cap_label = "",
  lt_continue = "\\\\footnotesize{continued on next page}",
  ...
)

\method{stable_long}{stobject}(data, ...)

\method{stable_long}{pmtable}(data, ...)
}
\arguments{
\item{data}{an object to render as a long table; this could be a \code{data.frame},
a \code{pmtable} object or an \code{stobject}; when passing in a \code{data.frame}, the data
should be filtered or subset so that \code{data} contains exactly the rows (and
columns) to be processed; pmtables will not add or remove rows prior to
processing \code{data}}

\item{...}{passed to \code{\link[=stable]{stable()}}}

\item{note_config}{a \code{\link[=noteconf]{noteconf()}} object used to configure how table notes
are displayed; see also \code{\link[=st_noteconf]{st_noteconf()}}.}

\item{inspect}{fixed to \code{TRUE} and passed to \code{\link[=stable]{stable()}}}

\item{lt_cap_macro}{the name of a macro that will hold caption text; to not
lead with \verb{\\\\} - this will be added for you}

\item{lt_cap_text}{full caption text, appearing where the table is rendered}

\item{lt_cap_short}{short caption text, appearing in the list of tables}

\item{lt_cap_label}{table label for use in latex document}

\item{lt_continue}{longtable continuation message}
}
\value{
A character vector with the TeX code for the table with \code{class}
attribute set to \code{stable_long} and \code{stable}.
}
\description{
Use this function to allow your table to span multiple pages, with a
"to be continued" statement at the bottom of each page. There are important
differences between this \code{longtable} environment and the \code{tabular}
environment that is used to generate tables from \code{\link[=stable]{stable()}}. See the
\code{details} section for more information.
}
\details{
To create \code{longtable} output, \code{pmtables} first passes the data frame
through \code{\link[=stable]{stable()}} and then modifies the output to create a table in
\code{longtable} environment. The \code{...} arguments to \code{\link[=stable_long]{stable_long()}} are passed
to \code{\link[=stable]{stable()}} and can be used to configure the table. One important difference
between \code{tablular} and \code{longtable} environments is that captions need to
get inserted \strong{inside} the \code{longtable} environment; this is why you see
several additional arguments for \code{\link[=stable_long]{stable_long()}}.

If you have panels in your table, the default is to prevent page breaks
right after the panel title row using the \verb{\\\\*} command in the \code{longtable}
package. This shouldn't need to be changed by the user, but if needed this
can be suppressed by adding \code{nopagebreak = FALSE} when calling \code{\link[=as.panel]{as.panel()}}
or \code{\link[=rowpanel]{rowpanel()}}.
}
\examples{
stable_long(stdata())

}
