% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{stable_save_image}
\alias{stable_save_image}
\title{Render and save a table in png or pdf format}
\usage{
stable_save_image(
  x,
  file = attr(x, "stable_file"),
  stem = NULL,
  format = c("png", "pdf"),
  dir = getOption("pmtables.dir"),
  build_dir = tempdir(),
  ...
)
}
\arguments{
\item{x}{an stable object.}

\item{file}{the final output file name; if \code{file} has \code{.tex} extension, the
extension will be replaced with the value in \code{format}.}

\item{stem}{used to form the final output file name; this argument will
override what is provided in \code{file}.}

\item{format}{the output file format.}

\item{dir}{the final output directory.}

\item{build_dir}{the directory where the image will be built.}

\item{...}{passed to \code{\link[=st_aspdf]{st_aspdf()}} or \code{\link[=st_aspng]{st_aspng()}}, depending on the value of
\code{format}.}
}
\description{
Render and save a table in png or pdf format
}
\details{
The table will be built / rendered in \code{tempdir()}.
}
\examples{
tab <- stable(stdata(), output_file = "foo.tex")

\dontrun{
  ans <- stable_save_image(tab format = "pdf")
  basename(ans)
  st_image_show(ans)
}

}
\seealso{
\code{\link[=st_as_image]{st_as_image()}}
}
