% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{[.pt_opts}
\alias{[.pt_opts}
\title{Extract multiple pmtables options}
\usage{
\method{[}{pt_opts}(x, i, ..., exact = TRUE)
}
\arguments{
\item{x}{the \code{pt_opts} object}

\item{i}{a character option name to extract}

\item{...}{other option names to extract}

\item{exact}{not used}
}
\description{
Extract multiple pmtables options
}
\examples{
pt["id_col", "dv_col"]

}
\keyword{internal}
