% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics-table.R
\name{sum_cont}
\alias{sum_cont}
\title{Default summary function in pt_demographics for continuous variables}
\usage{
sum_cont(value = seq(1, 5))
}
\arguments{
\item{value}{a vector or column to summarize}
}
\value{
A tibble with one row and two columns named \code{mean (sd)} and \code{min-max}
}
\description{
Default summary function in pt_demographics for continuous variables
}
\examples{
pmtables:::sum_cont(value = seq(1,5))

}
\keyword{internal}
