% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sumrow.R
\name{sumrow}
\alias{sumrow}
\title{Identify and style summary rows}
\usage{
sumrow(
  rows,
  col = 1,
  hline = TRUE,
  bold = FALSE,
  it = FALSE,
  blank = NULL,
  label = NULL
)
}
\arguments{
\item{rows}{integer or logical vector of rows that are summary rows}

\item{col}{integer or character (name) column location that labels the
summary row}

\item{hline}{logical; if \code{TRUE}, a horizontal line will be placed above the summary
row}

\item{bold}{logical; if \code{TRUE}, then the cell(s) identified by \code{rows} and
\code{col} will be rendered in bold font}

\item{it}{logical; if \code{TRUE}, then the cell(s) identified by \code{rows} and
\code{col} will be rendered in italic font}

\item{blank}{integer column positions in the summary row(s) to be made blank}

\item{label}{character label to replace text in cell(s) marked by \code{row(s)}
and `col}
}
\description{
Identify and style summary rows
}
