% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-clear_reps.R
\name{tab_clear_reps}
\alias{tab_clear_reps}
\title{Clear repeated values in a column}
\usage{
tab_clear_reps(
  data = NULL,
  clear_reps = NULL,
  panel = rowpanel(NULL),
  clear_grouped_reps = NULL,
  ...
)
}
\arguments{
\item{data}{a data.frame to convert to tabular table; see also \code{\link[=st_new]{st_new()}}}

\item{clear_reps}{character vector of column names where duplicate values will
be made blank (overwritten with \code{""}); see also \code{\link[=st_clear_reps]{st_clear_reps()}}}

\item{panel}{character column name to use to section the table; sections will
be created from unique values of \code{data[[panel]]}; see also \code{\link[=st_panel]{st_panel()}}}

\item{clear_grouped_reps}{passed to \code{\link[=clear_grouped_values]{clear_grouped_values()}} as \code{cols}}

\item{...}{not used}
}
\description{
Clear repeated values in a column
}
