% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sizes.R
\name{tab_size}
\alias{tab_size}
\title{Adjustments for row padding, column padding, and font size}
\usage{
tab_size(row = 1.4, col = 5, font = NULL)
}
\arguments{
\item{row}{relative increase or decrease spacing between rows; use
\verb{row_space > <default>} to increase; ; see also \code{\link[=st_space]{st_space()}}}

\item{col}{absolute column spacing amount (\code{pt}); see also \code{\link[=st_space]{st_space()}}}

\item{font}{for the table (e.g. \code{normalsize}, \code{small}, \code{scriptsize}, etc)}
}
\description{
Adjustments for row padding, column padding, and font size
}
