% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-span.R
\name{tab_spanners}
\alias{tab_spanners}
\title{Create groups of columns with spanners}
\usage{
tab_spanners(
  data,
  cols = NULL,
  span = NULL,
  span_split = NULL,
  span_title_break = "...",
  sizes = tab_size(),
  ...
)
}
\arguments{
\item{data}{a data.frame to convert to tabular table; the user should filter
or subset so that \code{data} contains exactly the rows (and columns) to be
processed; pmtables will not add or remove rows prior to processing \code{data};
see also \code{\link[=st_new]{st_new()}}}

\item{cols}{a character vector of column names}

\item{span}{a list of objects created with \code{\link[=colgroup]{colgroup()}}; ; see also \code{\link[=st_span]{st_span()}}}

\item{span_split}{not implemented at this time; ; see also \code{\link[=st_span_split]{st_span_split()}}}

\item{span_title_break}{a character sequence indicating where to split the
title across multiple lines}

\item{sizes}{an object returned from \code{\link[=tab_size]{tab_size()}}}

\item{...}{not used}
}
\description{
Create groups of columns with spanners
}
