% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-utils.R
\name{tex_bold}
\alias{tex_bold}
\alias{tex_it}
\title{Make bold}
\usage{
tex_bold(x, pattern = "*")

tex_it(x, pattern = "*")
}
\arguments{
\item{x}{a string}

\item{pattern}{a regular expression}
}
\description{
Make bold
}
