% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-stable.R
\name{triage_data}
\alias{triage_data}
\title{Get data ready for processing with s table}
\usage{
triage_data(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
Get data ready for processing with s table
}
\details{
\enumerate{
\item \code{data} must be a data frame
\item \code{data} is ungrouped with \code{\link[dplyr:group_by]{dplyr::ungroup()}}
\item factor columns in \code{data} are converted to character
}
}
