% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{update_abbrev}
\alias{update_abbrev}
\title{Update the abbreviation for a glossary entry}
\usage{
update_abbrev(x, ...)
}
\arguments{
\item{x}{a glossary object.}

\item{...}{\verb{<label> = <new abbreviation>} pairs; if just the
\verb{<label>} is passed, it will be used as the new update
abbreviation value.}
}
\description{
Update the abbreviation for a glossary entry
}
\examples{
x <- as_glossary(list(ka = "absorption rate constant"))

x <- update_abbrev(x, ka = "\\\\ensuremath{k_a}")

x$ka

x <- update_abbrev(x, ka)

x$ka

}
\seealso{
\code{\link[=as_glossary]{as_glossary()}}
}
