% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics-table.R
\name{validate_dem_fun}
\alias{validate_dem_fun}
\title{Check if summary function evaluates to the correct format}
\usage{
validate_dem_fun(fun)
}
\arguments{
\item{fun}{The summary function to use for summarizing the continuous data in \code{\link[=pt_demographics]{pt_demographics()}}}
}
\description{
Check if summary function evaluates to the correct format
}
\keyword{internal}
