% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-as-df.R
\name{yaml_as_df}
\alias{yaml_as_df}
\title{Read yaml input into data frame}
\usage{
yaml_as_df(path)
}
\arguments{
\item{path}{to the yaml source file}
}
\description{
Read yaml input into data frame
}
\examples{

path <- system.file("table.yml", package = "pmtables")

yaml_as_df(path)

}
