% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{analysis1}
\alias{analysis1}
\title{analysis1 data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4680 rows and 36 columns.
}
\usage{
analysis1
}
\description{
analysis1 data set
}
\keyword{datasets}
