% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption.R
\name{as.caption}
\alias{as.caption}
\title{Create caption}
\usage{
as.caption(
  x,
  short = NULL,
  short_repeat = TRUE,
  short_sep = NULL,
  write = FALSE
)
}
\arguments{
\item{x}{caption text.}

\item{short}{an abbreviated form of the caption to be used in a list of
tables.}

\item{short_repeat}{logical; if a short caption is provided, it will be
repeated at the start of the main caption.}

\item{short_sep}{a character sequence used to separate the short title
with the main caption when a short title is specified.}

\item{write}{logical; if \code{TRUE}, caption will be written to output file
when \code{\link[=stable_save]{stable_save()}} is called; this argument does not apply to
\code{stable_long} objects whose captions are \emph{always} written to output the
output file.}
}
\value{
A character vector of (main) caption text is returned with
attributes \code{short} and \code{write}.
}
\description{
Create caption
}
\examples{

as.caption("[Short]. Main title text.")
as.caption("Main title text.", short = "Short")

}
