% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-panel.R
\name{as.panel}
\alias{as.panel}
\alias{as.panel.rowpanel}
\alias{as.panel.character}
\alias{as.panel.quosures}
\title{Coerce object to rowpanel}
\usage{
as.panel(x, ...)

\method{as.panel}{rowpanel}(x, ...)

\method{as.panel}{character}(x, ...)

\method{as.panel}{quosures}(x, ...)
}
\arguments{
\item{x}{object to coerce}

\item{...}{arguments passed to \code{\link[=rowpanel]{rowpanel()}}}
}
\description{
Coerce object to rowpanel
}
\examples{
as.panel("STUDY", prefix = "Study: ")

}
