% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{as_glossary}
\alias{as_glossary}
\title{Coerce a named list to glossary}
\usage{
as_glossary(x, ...)
}
\arguments{
\item{x}{a named list.}

\item{...}{unquoted \verb{<label> = <new abbreviation>} pairs.}
}
\description{
Coerce a named list to glossary
}
\examples{

l <- list(VPC = "visual predictive check", tz = "timezone")

as_glossary(l)

as_glossary(l, tz = "tzone")


}
\seealso{
\code{\link[=read_glossary]{read_glossary()}}

\code{\link[=update_abbrev]{update_abbrev()}}
}
