% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-stable.R
\name{as_stable}
\alias{as_stable}
\alias{as_stable.pmtable}
\alias{as_stable.stable}
\alias{as_stable.stobject}
\title{Create stable from pmtable}
\usage{
as_stable(x, ...)

\method{as_stable}{pmtable}(x, ..., long = FALSE, wrap = FALSE, wrapw = FALSE, con = NULL)

\method{as_stable}{stable}(x, ...)

\method{as_stable}{stobject}(x, ...)
}
\arguments{
\item{x}{object to convert to stable}

\item{...}{for the \code{pmtable} method, these are extra named arguments to pass
to \code{\link[=stable]{stable()}}}

\item{long}{if \code{TRUE}, render with \code{\link[=stable_long]{stable_long()}} to create a longtable;
otherwise, by default process with \code{\link[=stable]{stable()}}}

\item{wrap}{if \code{TRUE}, the stable output will be wrapped in a latex table
environment using \code{\link[=st_wrap]{st_wrap()}}}

\item{wrapw}{if \code{TRUE}, the stable output will be wrapped in a latex table
environment and the output will be written to \code{\link[=stdout]{stdout()}}; use this along with
\code{results = "asis"} when rendering tables with \code{\link[rmarkdown:render]{rmarkdown::render()}}}

\item{con}{passed to \code{\link[=st_wrap]{st_wrap()}}; used when \code{wrap} is \code{TRUE}}
}
\description{
Create stable from pmtable
}
\keyword{internal}
