% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{cat_data}
\alias{cat_data}
\title{Summarize categorical data}
\usage{
cat_data(
  data,
  cols,
  by = ".total",
  summarize_all = TRUE,
  all_name = "All",
  wide = FALSE,
  nby = NULL,
  preshape = FALSE
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{character vector of column names for summary}

\item{summarize_all}{if \code{TRUE} then a complete data summary will be appended
to the bottom of the table}

\item{all_name}{a name to use for the complete data summary}

\item{wide}{if \code{TRUE} covariates are rendered west to east; if \code{FALSE} then
they are rendered north to south}

\item{nby}{number of unique levels for the \code{by} variable}

\item{preshape}{if \code{TRUE}, returns summarized data prior to reshaping;
this is intended for internal use}
}
\description{
Summarize categorical data
}
\examples{
data <- pmtables:::data("id")

cat_data(data, cols = c(SEX = "SEXf", RF = "RFf"), by = "STUDYf")

}
