% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{cat_data}
\alias{cat_data}
\title{Summarize categorical data}
\usage{
cat_data(
  data,
  cols,
  by = ".total",
  panel = by,
  summarize_all = TRUE,
  all_name = "All",
  wide = FALSE,
  nby = NULL,
  preshape = FALSE
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{by}{a grouping variable; may be character vector or quosure}

\item{panel}{data set column name to stratify the summary}

\item{summarize_all}{logical indicating whether or not to include a summary
of the full data in the output}

\item{all_name}{label for full data summary}

\item{wide}{\code{logical}; if \code{TRUE}, data frame will be returned in wide format;
if \code{FALSE}, it will be returned in \code{long} format}

\item{nby}{number of unique levels for the \code{by} variable}

\item{preshape}{if \code{TRUE}, returns summarized data prior to reshaping;
this is intended for internal use}
}
\description{
Summarize categorical data
}
\examples{

cat_data(pmt_first, cols = c(SEX = "SEXf", RF = "RFf"), by = "STUDYf")

}
