% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-align.R
\name{col_ragged}
\alias{col_ragged}
\title{Format a table cell with fixed width}
\usage{
col_ragged(
  size,
  unit = "cm",
  ragged = c("right", "left"),
  coltype = c("p", "m", "b")
)
}
\arguments{
\item{size}{the width of the cell}

\item{unit}{to go with \code{size}}

\item{ragged}{use \code{right} to left justify and \code{left} to right justify}

\item{type}{column type}
}
\description{
Format a table cell with fixed width
}
