% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-span.R
\name{colgroup}
\alias{colgroup}
\alias{is.colgroup}
\title{Set spans to group columns}
\usage{
colgroup(title, vars = c(), level = 1, sep = NULL)

is.colgroup(x)
}
\arguments{
\item{title}{label for groups of columns}

\item{vars}{tidyselect specification of variables}

\item{level}{relative position for the grouping spanner; level 0 is the
column names; level 1 is one step away (up) from the column names, etc}

\item{sep}{not implemented yet}

\item{x}{an R object}
}
\value{
an object with class \code{colgroup}
}
\description{
Set spans to group columns
}
