% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-span.R
\name{colgroup}
\alias{colgroup}
\alias{colsplit}
\alias{is.colgroup}
\title{Set spans to group columns}
\usage{
colgroup(title = NULL, vars = c(), level = 1, sep = ".", split = FALSE)

colsplit(title, level = 1, split = TRUE, sep = ".", ...)

is.colgroup(x)
}
\arguments{
\item{title}{label for groups of columns}

\item{vars}{tidyselect specification of variables}

\item{level}{relative position for the grouping spanner; level 0 is the
column names; level 1 is one step away (up) from the column names, etc}

\item{sep}{character; the separator used for finding column groupings}

\item{split}{logical; if \code{TRUE} column groupings will be determined by
splitting columns names on a separator}

\item{...}{arguments passed to \link{colgroup}}

\item{x}{an R object}
}
\value{
an object with class \code{colgroup}
}
\description{
Set spans to group columns
}
