% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-span.R
\name{colgroup}
\alias{colgroup}
\alias{as.span}
\alias{colsplit}
\alias{is.colgroup}
\alias{is.colsplit}
\title{Set spans to group columns}
\usage{
colgroup(
  title = NULL,
  vars = c(),
  level = 1,
  sep = ".",
  split = FALSE,
  title_side = c(".l", ".r")
)

as.span(
  title = NULL,
  vars = c(),
  level = 1,
  sep = ".",
  split = FALSE,
  title_side = c(".l", ".r")
)

colsplit(
  sep,
  level = 1,
  split = TRUE,
  title = NULL,
  title_side = c(".l", ".r")
)

is.colgroup(x)

is.colsplit(x)
}
\arguments{
\item{title}{label for groups of columns}

\item{vars}{tidyselect specification of variables}

\item{level}{relative position for the grouping spanner; level 0 is the
column names; level 1 is one step away (up) from the column names, etc}

\item{sep}{character; the separator used for finding column groupings}

\item{split}{logical; if \code{TRUE} column groupings will be determined by
splitting columns names on a separator}

\item{title_side}{which side of the split should be taken as the \code{title}?
defaults to \code{.left} but can also take the \code{.right} side of the splits}

\item{x}{an R object}
}
\value{
an object with class \code{colgroup}
}
\description{
Set spans to group columns
}
