% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-align.R
\name{cols_align}
\alias{cols_align}
\alias{cols_center}
\alias{cols_left}
\alias{cols_right}
\title{Set table column alignment}
\usage{
cols_align(
  .default = "l",
  ...,
  .r = NULL,
  .c = NULL,
  .l = NULL,
  .coltype = "p",
  .outer = c("none", "l", "r", "lr"),
  .complete = NULL
)

cols_center(...)

cols_left(...)

cols_right(...)
}
\arguments{
\item{.default}{the default column alignment}

\item{...}{exceptions to use from \code{.default}; each argument should be named
by a column in the data set; values should be either "l", "c", or "r";
for example: \code{WT = "l"}}

\item{.r}{column names as character vector or comma separated string
to align right; for example: \code{.r = "WT,AGE"}}

\item{.c}{column names as character vector or comma separated string
to center; ; for example: \code{.c = "WT,AGE"}}

\item{.l}{column names as character vector or comma separated string
to align left; ; for example: \code{.l = "WT,AGE"}}

\item{.coltype}{should be p, m, or b}

\item{.outer}{force the left-most column to the left (\code{l}), or the right-most
column on the right (\code{r}), or both (\code{lr})}

\item{.complete}{not used}
}
\description{
Set table column alignment
}
