% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-align.R
\name{cols_align}
\alias{cols_align}
\alias{cols_center}
\alias{cols_left}
\alias{cols_right}
\title{Set table column alignment}
\usage{
cols_align(.default = c("l", "c", "r"), ..., .complete = NULL)

cols_center(...)

cols_left(...)

cols_right(...)
}
\arguments{
\item{.default}{the default column alignment}

\item{...}{exceptions to use from \code{default}; each argument should be named
by a column in the data set; values should be either "l", "c", or "r"}

\item{.complete}{not used}
}
\description{
Set table column alignment
}
