% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{cont_table_data}
\alias{cont_table_data}
\title{Create continuous summary data frame}
\usage{
cont_table_data(
  data,
  cols,
  by = ".total",
  panel = by,
  wide = FALSE,
  all_name = "all",
  digits = new_digits(),
  id_col = "ID",
  fun = df_sum_2
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{by}{grouping variable name}

\item{panel}{paneling variable name}

\item{wide}{\code{logical}; if \code{TRUE}, output will be wide; output will be long
otherwise}

\item{all_name}{label for full data summary}

\item{digits}{named list specifing \code{digits} argument for \code{digit_fun}}

\item{id_col}{the ID column name}

\item{fun}{continuous data summary function}
}
\description{
Create continuous summary data frame
}
