% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{cont_table_data}
\alias{cont_table_data}
\title{Create continuous summary data frame}
\usage{
cont_table_data(
  data,
  cols,
  by = ".total",
  wide = FALSE,
  all_name = "all",
  digits = NULL,
  fun = df_sum_2
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{by}{grouping variable name}

\item{wide}{if \code{TRUE} covariates are rendered west to east; if \code{FALSE} then
they are rendered north to south}

\item{all_name}{label for full data summary}

\item{digits}{named list specifing \code{digits} argument for \code{digit_fun}}

\item{fun}{continuous data summary function}
}
\description{
Create continuous summary data frame
}
