% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{cont_table_data}
\alias{cont_table_data}
\title{Create continuous summary data frame}
\usage{
cont_table_data(
  data,
  cols,
  by = ".total",
  panel = by,
  wide = FALSE,
  all_name = "all",
  digits = new_digits(),
  id_col = "ID",
  na_fill = "--",
  fun = cont_long_fun
)
}
\arguments{
\item{data}{the data frame to summarize; the user should filter or subset
so that \code{data} contains exactly the records to be summarized; pmtables will
not add or remove rows prior to summarizing \code{data}}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{by}{grouping variable name}

\item{panel}{paneling variable name}

\item{wide}{\code{logical}; if \code{TRUE}, output will be wide; output will be long
otherwise}

\item{all_name}{label for full data summary}

\item{digits}{named list specifying \code{digits} argument for \code{digit_fun}}

\item{id_col}{the ID column name}

\item{na_fill}{value to fill with when all values in the summary are missing}

\item{fun}{continuous data summary function}
}
\description{
Create continuous summary data frame
}
