% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-functions.R
\name{cont_wide_fun}
\alias{cont_wide_fun}
\title{Function for continuous wide summaries}
\usage{
cont_wide_fun(value, digit_fun = sig, id = NULL, digits = 3, ...)
}
\arguments{
\item{value}{the data to summarize}

\item{digit_fun}{a function to format digits in the summaries}

\item{id}{a vector of subjet IDs; same length as \code{value}}

\item{digits}{the number of digits in the summary; the current implementation
passes \code{digits} to \code{digit_fun()}}

\item{...}{not used}
}
\description{
Function for continuous wide summaries
}
\examples{
pmtables:::cont_wide_fun(rnorm(100))

}
\keyword{internal}
