% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{data_inventory_chunk}
\alias{data_inventory_chunk}
\title{Create a data inventory summary for a data chunk}
\usage{
data_inventory_chunk(
  data,
  by,
  panel = by,
  stacked = FALSE,
  tot = FALSE,
  all_name = "all",
  dv_col = "DV",
  bq_col = "BQL",
  id_col = "ID",
  ...
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{by}{the outer grouping variable; may be character or quosure}

\item{panel}{the panel grouping variable; may be character or quosure}

\item{stacked}{if \code{TRUE}, then independent summaries are created by \code{outer}
and included in a single table (see examples)}

\item{tot}{logical indicating if a summary row should be included}

\item{all_name}{a name to use for the complete data summary}

\item{dv_col}{character name of \code{DV} column}

\item{bq_col}{character name of \code{BQL} column}

\item{id_col}{character name of \code{ID} column}

\item{...}{used to absorb other arguments; not used}
}
\description{
Create a data inventory summary for a data chunk
}
