% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{data_inventory_data}
\alias{data_inventory_data}
\title{Create a summary of endpoint data}
\usage{
data_inventory_data(
  data,
  by,
  panel = by,
  all_name = "all",
  summarize_all = TRUE,
  stacked = FALSE,
  ...
)
}
\arguments{
\item{data}{the data frame to summarize; the user should filter or subset
so that \code{data} contains exactly the records to be summarized; pmtables will
not add or remove rows prior to summarizing \code{data}}

\item{by}{The outer grouping variable; may be character or quosure.}

\item{panel}{The panel grouping variable; may be character or quosure.}

\item{all_name}{a name to use for the complete data summary}

\item{summarize_all}{if \code{TRUE} then a complete data summary will be
appended to the bottom of the table when \code{stacked} is \code{FALSE}.}

\item{stacked}{If \code{TRUE}, then independent summaries are created by \code{outer}
and included in a single table (see examples).}

\item{...}{passed to subsequent summary functions}
}
\description{
Create a summary of endpoint data
}
