% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{data_inventory_data}
\alias{data_inventory_data}
\title{Create a summary of endpoint data}
\usage{
data_inventory_data(
  data,
  outer,
  inner = outer,
  all_name = "all",
  stacked = FALSE,
  ...
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{outer}{the outer grouping variable; may be character or quosure}

\item{inner}{the inner grouping variable; may be character or quosure}

\item{all_name}{a name to use for the complete data summary}

\item{stacked}{if \code{TRUE}, then independent summaries are created by \code{outer}
and included in a single table (see examples)}

\item{...}{passed to subsequent summary functions}
}
\description{
Create a summary of endpoint data
}
