% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics-table.R
\name{dem_cont_fun}
\alias{dem_cont_fun}
\title{Default summary function in pt_demographics for continuous variables}
\usage{
dem_cont_fun(
  value = seq(1, 5),
  name = "",
  ...,
  fmt = sig,
  digits = 3,
  maxex = 5
)
}
\arguments{
\item{value}{a vector or column to summarize}

\item{name}{the name of the variable currently being summarized}

\item{...}{other arguments passed to \code{fmt}}

\item{fmt}{a function to format the numbers in the continuous data summary}

\item{digits}{passed to \code{fmt}}

\item{maxex}{passed to \code{fmt}}
}
\value{
A tibble with one row and three columns:
\enumerate{
\item \code{Mean (SD)} the mean and standard deviation of \code{value}
\item \code{Min / Max} the range of the \code{value}
\item \code{Missing} the number of missing \code{values}
}
}
\description{
Default summary function in pt_demographics for continuous variables
}
\details{
The summary function should have these arguments:
\itemize{
\item \code{value} the data to summarize
\item \code{name} the name of the data column being summarized; length 1
\item \code{...} absorbs other arguments that might get passed
}

These are the important arguments; \code{name} is not required as a formal
argument but \code{...} is.  The idea is that the custom summary function
receives the data to be summarized, the name of that data, and other
arguments. Since you are in charge of the function, you can make decisions
about both summaries and the formatting of those summaries inside the
function you are maintaining. Notice that \code{dem_cont_fun} also contains
arguments \code{fmt} (a formatting function) and \code{digits} and \code{maxex} (arguments
to be passed to \code{fmt}). This means that you can have your summary function
call the default summary function but with different formatting parameters.

The summary function should return a data frame, with summary statistics
in the columns.
}
\examples{
pmtables:::dem_cont_fun(value = seq(1,7), digits = 2)

}
\keyword{internal}
