% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics-table.R
\name{dem_cont_fun}
\alias{dem_cont_fun}
\title{Default summary function in pt_demographics for continuous variables}
\usage{
dem_cont_fun(value = seq(1, 5), digits = 3)
}
\arguments{
\item{value}{a vector or column to summarize}

\item{digits}{the number of digits in the summary; the current implementation
passes digits to \code{\link[=sig]{sig()}}}
}
\value{
A tibble with one row and three columns:
\enumerate{
\item \code{mean (sd)} the mean and standard deviation of \code{value}
\item \code{min-max} the range of the \code{value}
\item \code{non-missing} the number of non-missing \code{values}
}
}
\description{
Default summary function in pt_demographics for continuous variables
}
\details{
The summary function must return a data frame (or tibble) with a single
row and data summaries in the columns. Results will be coerced to
character.
}
\examples{
pmtables:::dem_cont_fun(value = seq(1,7), digits = 2)

}
\keyword{internal}
