% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{glossary_notes}
\alias{glossary_notes}
\alias{glossary_notes.character}
\alias{glossary_notes.list}
\alias{glossary_notes.glossary}
\title{Return formatted notes from a tex glossary file or glossary object}
\usage{
glossary_notes(x, ...)

\method{glossary_notes}{character}(x, ...)

\method{glossary_notes}{list}(x, ...)

\method{glossary_notes}{glossary}(x, ..., sep = ": ", collapse = "; ", labels = NULL)
}
\arguments{
\item{x}{path to a tex glossary file, a glossary object, or a list that can
be coerced to a glossary object with \code{\link[=as_glossary]{as_glossary()}}.}

\item{...}{unquoted names matching those names in \code{glossary}.}

\item{sep}{character to separate name and value.}

\item{collapse}{a character used to collapse definitions into a
single string.}

\item{labels}{a character vector or comma-separates string of definition
labels to get appended to the end of names passed in \code{...}.}
}
\description{
Return formatted notes from a tex glossary file or glossary object
}
\examples{
file <- system.file("tex", "glossary.tex", package = "pmtables")

glossary_notes(file, WT, CRCL)

}
\seealso{
\code{\link[=st_notes_glo]{st_notes_glo()}}, \code{\link[=read_glossary]{read_glossary()}}
}
