% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{glossary_notes}
\alias{glossary_notes}
\title{Return formatted notes from a tex glossary file}
\usage{
glossary_notes(file, ..., sep = ": ", collapse = "; ", labels = NULL)
}
\arguments{
\item{file}{path to the tex glossary file.}

\item{...}{unquoted names matching those names in \code{glossary}.}

\item{sep}{character to separate name and value.}

\item{collapse}{a character used to collapse definitions into a
single string.}

\item{labels}{a character vector or comma-separates string of definition
labels to get appended to the end of names passed in \code{...}.}
}
\description{
Return formatted notes from a tex glossary file
}
\examples{
file <- system.file("tex", "glossary.tex", package = "pmtables")

glossary_notes(file, WT, CRCL)

}
\seealso{
\code{\link[=st_notes_glo]{st_notes_glo()}}, \code{\link[=read_glossary]{read_glossary()}}
}
