% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{pmt_first}
\alias{pmt_first}
\title{analysis1 data set - first record only}
\format{
A data frame with 160 rows and 36 variables
}
\usage{
pmt_first
}
\description{
This is the first record from each individual in the
\link{analysis1} data set.
}
\seealso{
\link{analysis1}, \link{pmt_pk}, \link{pmt_obs}
}
\keyword{datasets}
