% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{pmt_summarized}
\alias{pmt_summarized}
\title{analysis1 data set - summarized}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 9 columns.
}
\usage{
pmt_summarized
}
\description{
analysis1 data set - summarized
}
\keyword{datasets}
