% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_long}
\alias{pt_cat_long}
\title{Discrete data summary in long format}
\usage{
pt_cat_long(
  data,
  cols,
  span = ".total",
  all_name = "All Data",
  all_name_span = "All Groups",
  summarize_all = TRUE,
  table = NULL,
  by = NULL
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{span}{variable name for column spanner}

\item{all_name}{a name to use for the complete data summary}

\item{all_name_span}{table column name to use for data summaries across
levels of \code{span} if it is provided}

\item{summarize_all}{if \code{TRUE} then a complete data summary will be appended
to the bottom of the table}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{by}{use \code{span} argument instead}
}
\description{
Discrete data summary in long format
}
