% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_long}
\alias{pt_cat_long}
\alias{pt_cat_wide}
\title{Create categorical data summary tables}
\usage{
pt_cat_long(
  data,
  cols,
  span = by,
  by = ".total",
  all_name = "All Groups",
  summarize_all = TRUE,
  table = NULL
)

pt_cat_wide(
  data,
  cols,
  by = ".total",
  panel = by,
  table = NULL,
  all_name = "All data",
  summarize_all = TRUE
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{span}{variable name for column spanner}

\item{by}{variable name for grouping}

\item{all_name}{a name to use for the complete data summary}

\item{summarize_all}{if \code{TRUE} then a complete data summary will be appended
to the bottom of the table}

\item{table}{a named list for renaming data items}

\item{panel}{data set column name to stratify the summary}
}
\description{
Create categorical data summary tables
}
