% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_study}
\alias{pt_cat_study}
\title{Discrete covariate table by study}
\usage{
pt_cat_study(
  data,
  cols,
  study_col = vars(`Study ID` = all_of("STUDY")),
  summarize_all = TRUE,
  all_name = "All studies",
  table = NULL,
  wide = FALSE
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{study_col}{character name of the data set column containing the study
identifier}

\item{summarize_all}{if \code{TRUE} then a complete data summary will be appended
to the bottom of the table}

\item{all_name}{label for full data summary}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{wide}{if \code{TRUE} the table will be rendered in a wide format}
}
\description{
Discrete covariate table by study
}
\examples{


ans <- pt_cat_study(pmt_first, cols = "SEXf,FORMf", study = "STUDYf")

ans <- pt_cat_study(pmt_first, cols = "FORMf", study = "STUDYf", wide = TRUE)

}
