% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{pt_cat_wide}
\alias{pt_cat_wide}
\title{Discrete data summary in long format}
\usage{
pt_cat_wide(
  data,
  cols,
  by = ".total",
  panel = by,
  table = NULL,
  all_name = "All data",
  summarize = c("bottom", "none"),
  drop = character(0)
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{by}{a grouping variable for the summary; may be given as character
vector or quosure}

\item{panel}{data set column name to stratify the summary}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{all_name}{a name to use for the complete data summary}

\item{summarize}{where to put an all-data summary; choose \code{none} to omit the
summary from the table}

\item{drop}{passed to \code{\link[=stable]{stable()}}}
}
\description{
Discrete data summary in long format
}
