% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{pt_cont_long}
\alias{pt_cont_long}
\title{Continuous data summary in long format}
\usage{
pt_cont_long(
  data,
  cols,
  panel = ".total",
  panel_invert = FALSE,
  table = NULL,
  units = NULL,
  digits = new_digits(),
  summarize_all = TRUE,
  all_name = "All data",
  fun = cont_long_fun,
  id_col = "ID"
)
}
\arguments{
\item{data}{the data frame to summarize}

\item{cols}{the columns to summarize; may be character vector or quosure}

\item{panel}{data set column name to stratify the summary}

\item{panel_invert}{if \code{TRUE} and \code{panel} is passed, the table will be
paneled by the covariate name and the \code{panel} variable will be repeated
underneath the covariate name; this allows the reader to see how one
covariate value varies within the different levels of \code{panel}}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{units}{a named list to use for unit lookup (see details and examples)}

\item{digits}{a \code{digits} object (see \code{\link[=new_digits]{new_digits()}})}

\item{summarize_all}{if \code{TRUE} then a complete data summary will be appended
to the bottom of the table}

\item{all_name}{a name to use for the complete data summary}

\item{fun}{the data summary function (see details)}

\item{id_col}{the ID column name}
}
\description{
Continuous data summary in long format
}
\details{
The summary function (\code{fun}) should take \code{value} as the first argument and
return a data frame or tibble with one row as many columns as you wish to
appear in the table. The function can also accept an \code{id} argument which is
a vector of \code{IDs} that is the same length as \code{value}. Be sure to include
\code{...} to the function signature as other arguments will be passed along.
Make sure your function completely formats the output ... it will appear in
the table as you return from this function. See \code{\link[=cont_long_fun]{cont_long_fun()}} for
details on the default implementation.
}
\examples{

ans <- pt_cont_long(pmt_first, cols = dplyr::vars(WT,ALB,CRCL))

ans <- pt_cont_long(pmt_first, cols = "WT,CRCL", panel = "SEXf")

}
