% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{pt_data_inventory_notes}
\alias{pt_data_inventory_notes}
\title{Return table notes for pt_data_inventory}
\usage{
pt_data_inventory_notes(note_add = NULL)
}
\arguments{
\item{note_add}{additional notes to include}
}
\description{
See \code{\link[=pt_data_inventory]{pt_data_inventory()}}.
}
