% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics-table.R
\name{pt_demographics}
\alias{pt_demographics}
\title{Summarize continuous and categorical data in long format}
\usage{
pt_demographics(
  data,
  cols_cont,
  cols_cat,
  span = NULL,
  units = NULL,
  stat_name = "Statistic",
  stat_width = 2,
  summarize_all = TRUE,
  all_name = "All data",
  fun = dem_cont_fun,
  notes = pt_demographics_notes(),
  paneled = TRUE
)
}
\arguments{
\item{data}{the data frame to summarize; the user should filter or subset
so that data contains exactly the records to be summarized; pmtables will not
add or remove rows prior to summarizing data}

\item{cols_cont}{the continuous data columns to summarize; this argument
may be specified as a character vector, comma-separated string or
quosure}

\item{cols_cat}{the categorical columns to summarize; this argument
may be specified as a character vector, comma-separated string or
quosure}

\item{span}{variable name for column spanner}

\item{units}{optional units for each summarized column; must be a named list
where the names correspond with continuous data columns in \code{data}}

\item{stat_name}{name of statistic column}

\item{stat_width}{width (in cm) of the statistic column}

\item{summarize_all}{logical; if \code{TRUE}, summaries across all \code{span}
levels will be appended to the right hand side of the table}

\item{all_name}{a character name for the all data summary invoked by
\code{summarize_all}}

\item{fun}{The summary function to use for summarizing the continuous
data; the default is \code{\link[=dem_cont_fun]{dem_cont_fun()}}. The result will be validated with
\code{\link[=validate_dem_fun]{validate_dem_fun()}}.}

\item{notes}{notes a character vector of notes to place under the table}

\item{paneled}{logical; if \code{TRUE}, the table will be paneled with the
covariate names; otherwise, the covariate names will appear as the left-most
column with non-repeating names cleared and separated with \code{hline} (see
examples).}
}
\value{
An object of class \code{pmtable}.

An object with class pmtable; see class-pmtable.
}
\description{
This function makes a single table from both continuous and categorical data.
}
\details{
The categorical data is summarized using \code{\link[=pt_cat_long]{pt_cat_long()}}.
The default summary function for continuous variables is \code{\link[=dem_cont_fun]{dem_cont_fun()}}.
Please review that documentation for details on the default summary for this
table.

If you wish to define your own function, please ensure the output is in the
same format. Any number of columns is acceptable.
}
\examples{

out <- pt_demographics(
  data = pmt_first,
  cols_cont = c(Age = "AGE", Weight = "WT"),
  cols_cat = c(Sex = "SEXf", Race = "ASIANf"),
  units = list(WT = "kg"),
  span = c(Study = "STUDYf")
)

out <- pt_demographics(
  data = pmt_first,
  cols_cont = "AGE,WT",
  cols_cat = "SEXf,ASIANf",
  paneled = FALSE,
  span = "FORMf"
)
tab <- stable(out)

pmtables:::pt_demographics_notes()

new_fun <- function(value = seq(1,5)) {
value <- value[!is.na(value)]
 tibble::tibble(
  `mean` = sig(mean(value)),
  `median` = sig(median(value)),
  `min-max` = paste0(sig(range(value)), collapse = " - ")
 )
}

out <- pt_demographics(
  data = pmt_first,
  cols_cont = "AGE,WT",
  cols_cat = "SEXf,ASIANf",
  fun = new_fun
)

pmtables:::dem_cont_fun(rnorm(20))
new_fun(rnorm(20))


}
\seealso{
\code{\link[=pt_cont_long]{pt_cont_long()}}, \code{\link[=pt_cat_long]{pt_cat_long()}}
}
