% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inventory_table.R
\name{pt_inventory_long}
\alias{pt_inventory_long}
\alias{pt_data_inventory_long}
\title{Data inventory by covariate}
\usage{
pt_inventory_long(
  data,
  cols,
  drop_miss = FALSE,
  table = NULL,
  summarise_all = TRUE,
  all_name = "All data",
  dv_col = "DV",
  bq_col = find_bq_col(data),
  id_col = "ID",
  level_width = NULL
)

pt_data_inventory_long(
  data,
  cols,
  drop_miss = FALSE,
  table = NULL,
  summarise_all = TRUE,
  all_name = "All data",
  dv_col = "DV",
  bq_col = find_bq_col(data),
  id_col = "ID",
  level_width = NULL
)
}
\arguments{
\item{data}{the data frame to summarize; the user should filter or subset
so that \code{data} contains exactly the records to be summarized; pmtables will
not add or remove rows prior to summarizing \code{data}}

\item{cols}{data columns containing discrete data items for grouped
data inventory summaries.}

\item{drop_miss}{If \code{TRUE}, then \code{MISS} will be dropped, but only when all
\code{MISS} values are equal to zero.}

\item{table}{a named list to use for renaming columns (see details and
examples)}

\item{all_name}{a name to use for the complete data summary}

\item{dv_col}{Character name of \code{DV} column.}

\item{bq_col}{Character name of \code{BQL} column; see \code{\link[=find_bq_col]{find_bq_col()}}.}

\item{id_col}{Character name of \code{ID} column.}

\item{level_width}{width of the \code{level} column, specified in \code{cm}.}
}
\description{
Data inventory by covariate
}
\examples{

tab <- pt_inventory_long(
  pmtables::pmt_first,
  cols = c("FORMf", "SEXf", "RFf")
)

tab$data

}
\seealso{
\code{\link[=pt_data_inventory]{pt_data_inventory()}}
}
