% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{pt_opts}
\alias{pt_opts}
\alias{pt_options}
\alias{pt}
\title{Set global plot options}
\format{
An object of class \code{pt_opts} (inherits from \code{environment}) of length 10.

An object of class \code{pt_opts} (inherits from \code{environment}) of length 10.
}
\usage{
pt_options(id_col = "ID", dv_col = "DV", bq_col = "BQL")

pt_opts

pt
}
\arguments{
\item{id_col}{ID column name}

\item{dv_col}{DV column name}

\item{bq_col}{BQL column name}
}
\description{
Set global plot options
}
\details{
\link{pt} and \link{pt_opts} both refer to the same environment.

Global options can
be set in the environment with \code{pt$set(name = value)}.  There is also
a \code{.list} argument to \code{pt$set} that allows you to pass in a named list of
options to set.

Values can be extracted with \code{pt$get("name")}.

Because it is an environment, the \code{$} operator can also be used to get and
set values (see examples).

Other methods in the environment include \code{pt$as.list()}, \code{pt$mget()},
\code{pt$reset()}.  \code{pt$self} refers to the environment itself. A list of
default settings can be obtained with \code{pt$defaults}.  Methods for the
\code{pt_opts} object include: \link{print.pt_opts}, \link{as.list.pt_opts},
\code{\link{$<-.pt_opts}}.
}
\examples{

pt$set(id_col = "USUBJID")

pt$id_col <- "ID"

pt$bq_col

pt$reset()

x <- pt$as.list()

\dontrun{
 defs <- pt$defaults
 defs$dv_col <- "dv"
 pt$set(.list = defs)
}

}
\keyword{datasets}
