% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{pt_opts}
\alias{pt_opts}
\alias{pt_options}
\title{Set global plot options}
\format{
An object of class \code{pt_opts} (inherits from \code{environment}) of length 17.
}
\usage{
pt_options(
  id_col = "ID",
  dv_col = "DV",
  bq_col = "BQL",
  units = NULL,
  table = NULL,
  panel.label.add = TRUE,
  cont.long.fun = pmtables:::df_sum_2,
  cont.wide.fun = pmtables:::str_sum_2,
  notes.sanitize = TRUE,
  digits = NULL
)

pt_opts
}
\arguments{
\item{id_col}{ID column name}

\item{dv_col}{DV column name}

\item{bq_col}{BQL column name}

\item{units}{named list of units}

\item{table}{named list of column name transformations}

\item{panel.label.add}{logical; if \code{TRUE}, panel variable name will be pasted
to the panel variable value}

\item{cont.long.fun}{data summary function for continuous / long summaries}

\item{cont.wide.fun}{data summary function for continuous / wide summaries}

\item{notes.sanitize}{if \code{TRUE}, should the notes be passed through the
sanitize function}

\item{digits}{a \code{digits} object}
}
\description{
Set global plot options
}
\details{
\code{pt_opts} is the options environment.

Global options can
be set in the environment with \code{pt_opts$set(name = value)}.  There is also
a \code{.list} argument to \code{pt_opts$set} that allows you to pass in a named list of
options to set.

Values can be extracted with \code{pt_opts$get("name")}.

Because it is an environment, the \code{$} operator can also be used to get and
set values (see examples).

Other methods in the environment include \code{pt_opts$as.list()}, \code{pt_opts$mget()},
\code{pt_opts$reset()}.  \code{pt_opts$self} refers to the environment itself. A list of
default settings can be obtained with \code{pt$defaults}.  Methods for the
\code{pt_opts} object include: \link{print.pt_opts}, \link{as.list.pt_opts},
\code{\link{$<-.pt_opts}}.
}
\examples{

pt_opts$set(id_col = "USUBJID")

pt_opts$id_col <- "ID"

pt_opts$bq_col

pt_opts$reset()

x <- pt_opts$as.list()

\dontrun{
 defs <- pt_opts$defaults
 defs$dv_col <- "dv"
 pt_opts$set(.list = defs)
}

}
\keyword{datasets}
