% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{read_glossary}
\alias{read_glossary}
\title{Read and parse a glossary file}
\usage{
read_glossary(file, format = guess_glo_fmt(file))
}
\arguments{
\item{file}{path to the tex or yaml glossary file.}

\item{format}{the glossary file format; could be either \code{"tex"} or \code{"yaml"}
and will be inferred by default from the extension of \code{file}.}
}
\value{
A named list of glossary definitions with class \code{tex_glossary} and
\code{glossary}. The list names are acronym entry labels (i.e., what you would
pass to \verb{\\gls\{\}} when writing a tex document).
}
\description{
Files may be formatted as TeX glossary file or
in yaml format suitable for reading using
\code{\link[=yaml_as_df]{yaml_as_df()}} (see details).
}
\details{
The glossary file is expected to contain acronym entries with the form

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\newacronym\{label\}\{abbreviation\}\{definition\}
}\if{html}{\out{</div>}}

or

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\newacronym[options]\{label\}\{abbreviation\}\{definition\}
}\if{html}{\out{</div>}}
}
\examples{
file <- system.file("glo", "glossary.tex", package = "pmtables")

x <- read_glossary(file)

names(x)

x$WT

x$SC

}
\seealso{
\code{\link[=as_glossary]{as_glossary()}}
}
