% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{read_glossary}
\alias{read_glossary}
\alias{simplify_abbrev}
\title{Read and parse a tex glossary file}
\usage{
read_glossary(file)

simplify_abbrev(x, ...)
}
\arguments{
\item{file}{path to the tex glossary file.}
}
\value{
A named list of glossary definitions with class \code{tex_glossary} and
\code{glossary}. The list names are acronym entry labels (i.e., what you would
pass to \verb{\\gls\{\}} when writing a tex document).
}
\description{
Read and parse a tex glossary file
}
\details{
The glossary file is expected to contain acronym entries with the form

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\newacronym\{label\}\{abbreviation\}\{definition\}
}\if{html}{\out{</div>}}

or

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\newacronym[options]\{label\}\{abbreviation\}\{definition\}
}\if{html}{\out{</div>}}
}
\examples{
file <- system.file("tex", "glossary.tex", package = "pmtables")

x <- read_glossary(file)

names(x)

x$WT

x$SC

}
\seealso{
\code{\link[=as_glossary]{as_glossary()}}
}
