% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-panel.R
\name{rowpanel}
\alias{rowpanel}
\alias{is.rowpanel}
\title{Create panel object}
\usage{
rowpanel(
  col = NULL,
  prefix = "",
  skip = ".panel.skip.",
  prefix_name = FALSE,
  prefix_skip = NULL,
  duplicates_ok = FALSE,
  bold = TRUE,
  it = FALSE,
  hline = TRUE,
  jut = 0,
  nopagebreak = TRUE
)

is.rowpanel(x)
}
\arguments{
\item{col}{name of column to be used for creating panels}

\item{prefix}{to be added to each panel title}

\item{skip}{regular expression for panels to skip}

\item{prefix_name}{\code{logical}; if \code{TRUE}, then the prefix will be derived by
the name associated with \code{col} input to the function}

\item{prefix_skip}{a regular expression for identifying panel titles where
the prefix won't be applied}

\item{duplicates_ok}{if \code{FALSE}, an error is generated if more than one
panel will have the same header}

\item{bold}{render panel title in bold font face}

\item{it}{render panel title in italic font face}

\item{hline}{logical indicating whether or not to draw an \code{hline} above
the panel row; the first panel row never receives an \code{hline}}

\item{jut}{amount (in TeX \code{ex} units) by which the panel headers are
outdented from the first column in the main table and header. Consider
using a value of \code{2} for a clear offset.}

\item{nopagebreak}{if \code{TRUE} (the default) then the page will not break
immediately after a panel title \emph{when creating a longtable}; this argument
does nothing if you are not generating a \code{longtable}. Set to \code{FALSE} to
prevent the \code{nopagebreak} command (implemented as \code{*}) in the longtable
output.}

\item{x}{an object to test}
}
\description{
Objects may also be coerced with \code{\link[=as.panel]{as.panel()}}
}
\seealso{
\code{\link[=as.panel]{as.panel()}}
}
