% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-panel.R
\name{rowpanel}
\alias{rowpanel}
\alias{is.rowpanel}
\title{Create panel object}
\usage{
rowpanel(
  col = NULL,
  prefix = "",
  prefix_name = FALSE,
  prefix_skip = NULL,
  duplicates_ok = FALSE,
  bold = TRUE,
  it = FALSE
)

is.rowpanel(x)
}
\arguments{
\item{col}{name of column to be used for creating panels}

\item{prefix}{to be added to each panel title}

\item{prefix_name}{\code{logical}; if \code{TRUE}, then the prefix will be derived by
the name associated with \code{col} input to the function}

\item{prefix_skip}{a regular expression for identifying panel titles where
the prefix won't be applied}

\item{duplicates_ok}{if \code{FALSE}, an error is generated if more than one
panel will have the same header}

\item{bold}{render panel title in bold font face}

\item{it}{render panel title in italic font face}

\item{x}{an object to test}
}
\description{
Objects may also be coerced with \code{\link[=as.panel]{as.panel()}}
}
\seealso{
\code{\link[=as.panel]{as.panel()}}
}
