% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sig}
\alias{sig}
\alias{digit1}
\alias{rnd}
\title{Format digits}
\usage{
sig(x, digits = 3, maxex = getOption("pmtables.maxex", NULL), ...)

digit1(x, ...)

rnd(x, digits = 0, ...)
}
\arguments{
\item{x}{\code{numeric}; value to manipulate.}

\item{digits}{\code{numeric}; number of significant digits.}

\item{maxex}{\code{numeric}; maximum number of significant
digits before moving to scientific notation.}

\item{...}{other arguments that are not used.}
}
\value{
A character vector of formatted values.
}
\description{
Use \code{\link[=sig]{sig()}} to set the number of significant digits; use \code{\link[=digit1]{digit1()}} to limit
to one digit. See examples.
}
\details{
When \code{x} is an integer, \code{x} is returned after coercing to character, without
further processing.
}
\examples{
sig(1.123455)
sig(0.123455)
sig(1.123455, digits = 5)
sig(1123, maxex = 3)
sig(1123, maxex = 4)

sig(1L)

digit1(1.234)
digit1(1321.123)

}
