% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sig}
\alias{sig}
\title{Significant Digits}
\usage{
sig(x, digits = 3, maxex = NULL)
}
\arguments{
\item{x}{numeric, value to manipulate}

\item{digits}{numeric, number of significant digits Default: 3}

\item{maxex}{numeric, maximum number of significant
digits before moving to scientific notation, Default: NULL}
}
\value{
character
}
\description{
Set significant digits according to metrum SOP.
}
\examples{
sig(1.123455)
sig(0.123455)
sig(1.123455,digits = 5)
sig(1123,maxex = 3)
sig(1123,maxex = 4)
}
