% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{st2article}
\alias{st2article}
\title{Preview tables in a TeX article}
\usage{
st2article(
  ...,
  .list = NULL,
  ntex = 1,
  stem = "article",
  output_dir = tempdir(),
  margin = NULL,
  template = NULL
)
}
\arguments{
\item{...}{stable objects}

\item{.list}{a list of stable objects}

\item{ntex}{number of times to build the pdf file}

\item{stem}{name for the article file, without extension}

\item{output_dir}{the output directory for the rendered pdf file}

\item{margin}{the horizontal margin in \code{cm}; default is 3 cm; the vertical
margin is fixed to 3 cm}

\item{template}{an optional template for rendering the article; this normally
shouldn't be used}
}
\description{
This is experimental. Pass in either \code{stable} or \code{stable_long} objects,
and \code{\link[=st2article]{st2article()}} will write the output to temporary file and render the
collection of tables in a pdf document without using Rmarkdown or pandoc.
The rendering is accomplished through system call to \code{pdflatex}. See
\code{details}.
}
\details{
A working tex distribution is required to run this function.It is important
to review the different latex dependencies in the document template.  It is
assumed that all of these dependencies are available. See \code{examples} below
for code to open the template for review.

This function requires \code{pdflatex} to be installed and in your path to build
the document.  Run \code{system2("pdflatex", "-v")} to see if \code{pdflatex} properly
installed.
}
\examples{
template_file <- system.file("article", "article.tex", package = "pmtables")
template_contents <- readLines(template_file)

\dontrun{
  tab <- stable(ptdata())
  pmtables:::st2article(tab)
}

}
