% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{st2article}
\alias{st2article}
\title{Preview tables in a TeX article}
\usage{
st2article(
  ...,
  .list = NULL,
  npdf = 1,
  stem = "article",
  output_dir = tempdir(),
  margin = NULL
)
}
\arguments{
\item{...}{stable objects}

\item{.list}{a list of stable objects}

\item{stem}{name for the article, without extension}

\item{margin}{the left and right margin in \code{cm}; default is 3 cm}

\item{ntex}{number of times to build the pdf file}
}
\description{
This is experimental
}
\details{
It is important to review the different latex dependencies in the
document template.  It is assumed that all of these dependencies are
available.
}
\examples{
template_file <- system.file("article", "article.tex", package = "pmtables")
template_contents <- readLines(template_file)

\dontrun{
  tab <- stable(ptdata())
  pmtables:::st2article(tab)
}

}
