% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{st2article}
\alias{st2article}
\title{Preview tables in a TeX article}
\usage{
st2article(..., .list = NULL, npdf = 1, stem = "article")
}
\arguments{
\item{...}{stable objects}

\item{.list}{a list of stable objects}

\item{npdf}{number of times to build the pdf file}

\item{stem}{name for the article, without extension}
}
\description{
This is experimental
}
