% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{st2doc}
\alias{st2doc}
\title{Render a table a pdf document}
\usage{
st2doc(text, preview = TRUE)
}
\arguments{
\item{text}{\code{\link[=stable]{stable()}} output}

\item{preview}{if \code{TRUE}, the rendered pdf file is opened using
\code{\link[fs:file_show]{fs::file_show()}}}
}
\description{
Render a table a pdf document
}
