% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{st2doc}
\alias{st2doc}
\title{Render a table a pdf document}
\usage{
st2doc(
  text,
  preview = TRUE,
  output_dir = tempdir(),
  output_file = "st2doc.pdf",
  landscape = is_lscape(text)
)
}
\arguments{
\item{text}{\code{\link[=stable]{stable()}} output}

\item{preview}{if \code{TRUE}, the rendered pdf file is opened using
\code{\link[fs:file_show]{fs::file_show()}}}

\item{output_dir}{passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}

\item{output_file}{passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}

\item{landscape}{passed to \code{\link[=st_wrap]{st_wrap()}}}
}
\value{
the \code{text} is returned invisibly
}
\description{
Render a table a pdf document
}
\details{
In order to render the table in the pdf document, the following
packages must be installed, regardless of the type of table
you are trying to render:
\enumerate{
\item \code{threeparttable}
\item \code{booktabs}
\item \code{array}
\item \code{longtable}
\item \code{mathpazo}
\item \code{pdflscape}
}
}
\examples{

\dontrun{
  library(dplyr)
  ptdata()  \%>\% stable() \%>\% st2doc()
}

template <- system.file("rmd", "st2doc.Rmd", package = "pmtables")
# cat(readLines(template), sep = "\n")

}
