% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_align}
\alias{st_align}
\alias{st_center}
\alias{st_left}
\alias{st_right}
\title{Add column alignment information to st object}
\usage{
st_align(x, ...)

st_center(x, ...)

st_left(x, ...)

st_right(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{named arguments passed to \code{\link[=cols_align]{cols_align()}}}
}
\description{
See the \code{align} argument to \code{\link[=stable]{stable()}}. Note: these functions
always replace the current alignment in total.
}
\examples{
library(dplyr)

ob <- st_new(ptdata())

ob \%>\% st_align(.default = 'l') \%>\% st_make()

ob \%>\% st_center(N = 'l') \%>\% st_make()

}
