% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_args}
\alias{st_args}
\title{Add other arguments to st object}
\usage{
st_args(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{named arguments to be passed to \code{\link[=stable]{stable()}}}
}
\description{
Add other arguments to st object
}
