% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_aspdf}
\alias{st_aspdf}
\alias{st2pdf}
\title{Render stable object to pdf file}
\usage{
st_aspdf(
  x,
  stem = "pmt-standalone-preview",
  dir = tempdir(),
  font = "helvetica",
  textwidth = getOption("pmtables.textwidth", 6.5),
  border = getOption("pmtables.image.border", "0.2cm 0.7cm"),
  ntex = 1
)

st2pdf(
  x,
  stem = "pmt-standalone-preview",
  dir = tempdir(),
  font = "helvetica",
  textwidth = getOption("pmtables.textwidth", 6.5),
  border = getOption("pmtables.image.border", "0.2cm 0.7cm"),
  ntex = 1
)
}
\arguments{
\item{x}{an stable object; this can be the result of calling \code{\link[=stable]{stable()}} or
\code{\link[=stable_long]{stable_long()}}.}

\item{stem}{used to build intermediate and output file names.}

\item{dir}{directory for building the pdf file.}

\item{font}{the font to use; alternative values include \code{roboto} and
\code{utopia}; passed to \code{\link[=st_to_standalone]{st_to_standalone()}}.}

\item{textwidth}{the page width (in inches) when building with \code{pdflatex};
passed to \code{\link[=st_to_standalone]{st_to_standalone()}}; see details.}

\item{border}{passed as an option to \code{standalone} latex output type; see
details.}

\item{ntex}{\strong{Deprecated}. This argument no longer has an effect. Underneath,
\code{latexmk} is used to run the underlying command (e.g., \code{pdflatex}) as many
times as required.}
}
\value{
A string containing the path to the rendered \code{pdf} file.
}
\description{
Create a "standalone" \code{pdf} snippet from an stable object using the
\code{pdflatex} utility. The resultant \code{pdf} file is saved on disk and the
relative path to the file is returned. \code{st2pdf()} is an alias to
\code{st_as_pdf()}.
}
\details{
The \code{pdf} file is built using \code{pdflatex} so this utility must be installed.

The \code{textwidth} argument is set to 6.5 inches by default to mimic a 8.5 x 11
page with 1 inch margins on the left and right. Setting \code{textwidth} sets the
length of the \verb{\\textwidth} latex macro to that value and also inserts an
invisible rule across the page with that width as well. This means for
skinny tables, there will be whitespace on the left and right, but the font
in the resultant images will be similar regardless of the width of the
table. To skip setting the latex \verb{\\textwidth} macro, pass \code{NULL}.

The \code{border} argument can be one, two or four space-separated elements, each
formatted as \code{"<number><unit>"} (e.g. "0.2cm"); pass one element to set the
same border on all sides; two elements to set the border on left/right
(first) and top/bottom (second); pass four elements to have separate borders
for the left, bottom, right and top (see the documentation for the
\code{standalone} latex package).
}
\examples{

# check that latexmk and pdflatex are installed
\dontrun{
Sys.which("latexmk")
Sys.which("pdflatex")
}

\dontrun{
tab <- stable(stdata())
st_aspdf(tab)
}

# the template for building the image
temp <- system.file("tex", "standalone-preview.tex", package = "pmtables")
cat(temp, sep = "\n")

}
\seealso{
\code{\link[=st_aspng]{st_aspng()}}, \code{\link[=st_as_image]{st_as_image()}}, \code{\link[=st_image_show]{st_image_show()}}
}
