% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_aspdf}
\alias{st_aspdf}
\title{Render stable object in pdf format}
\usage{
st_aspdf(x, stem = "pmt-standalone-preview", dir = tempdir())
}
\arguments{
\item{x}{an stable object.}

\item{stem}{for intermediate and output file names.}

\item{dir}{directory for building the image.}
}
\value{
A string containing the full path to the rendered pdf file.
}
\description{
Render stable object in pdf format
}
