% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_aspdf}
\alias{st_aspdf}
\title{Render stable object in pdf format}
\usage{
st_aspdf(x, stem = "pmt-standalone-preview", dir = tempdir(), font = "roboto")
}
\arguments{
\item{x}{an stable object.}

\item{stem}{for intermediate and output file names.}

\item{dir}{directory for building the image.}

\item{font}{the font to use; pass \code{roboto} for sans serif and \code{utopia}
for serif; only these two options are available at this time.}
}
\value{
A string containing the path to the rendered pdf file.
}
\description{
Render stable object in pdf format
}
\seealso{
\code{\link[=st_aspng]{st_aspng()}}, \code{\link[=st_as_image]{st_as_image()}}, \code{\link[=st_image_show]{st_image_show()}}
}
