% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_aspng}
\alias{st_aspng}
\title{Render stable object in png format}
\usage{
st_aspng(
  x,
  stem = "pmt-standalone-preview",
  dir = tempdir(),
  dpi = 200,
  font = "roboto"
)
}
\arguments{
\item{x}{an stable object.}

\item{stem}{for intermediate and output file names.}

\item{dir}{directory for building the image.}

\item{dpi}{dots per inch for the resulting \code{png} file.}

\item{font}{the font to use; alternative values include \code{roboto} and
\code{utopia}.}
}
\value{
A string containing the path to the rendered \code{png} file.
}
\description{
Create a standalone png snippet
using \code{latex} and \code{dvipng} from an stable object; both functions are required
to be installed for this to work. The resultant \code{png} file is saved on disk
and the relative path to the file is returned.
}
\seealso{
\code{\link[=st_aspdf]{st_aspdf()}}, \code{\link[=st_as_image]{st_as_image()}}, \code{\link[=st_image_show]{st_image_show()}}
}
