% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_aspng}
\alias{st_aspng}
\alias{st2png}
\title{Render stable object in png format}
\usage{
st_aspng(
  x,
  stem = "pmt-standalone-preview",
  dir = tempdir(),
  font = "helvetica",
  textwidth = getOption("pmtables.text.width", 6.5),
  border = getOption("pmtables.image.border", "0.2cm 0.7cm"),
  ntex = 1,
  dpi = 200
)

st2png(
  x,
  stem = "pmt-standalone-preview",
  dir = tempdir(),
  font = "helvetica",
  textwidth = getOption("pmtables.text.width", 6.5),
  border = getOption("pmtables.image.border", "0.2cm 0.7cm"),
  ntex = 1,
  dpi = 200
)
}
\arguments{
\item{x}{an stable object; this can be the result of calling \code{\link[=stable]{stable()}} or
\code{\link[=stable_long]{stable_long()}}.}

\item{stem}{used to build intermediate and output file names.}

\item{dir}{directory for building the pdf file.}

\item{font}{the font to use; alternative values include \code{roboto} and
\code{utopia}; passed to \code{\link[=st_to_standalone]{st_to_standalone()}}.}

\item{textwidth}{the page width (in inches) when building with \code{pdflatex};
passed to \code{\link[=st_to_standalone]{st_to_standalone()}}; see details.}

\item{border}{passed as an option to \code{standalone} latex output type; see
details.}

\item{ntex}{number of times to build the pdf file}

\item{dpi}{dots per inch for the resulting \code{png} file; used by \code{dvipng} when
converting \code{dvi} file to the final \code{png} result.}
}
\value{
A string containing the path to the rendered \code{png} file.
}
\description{
Create a standalone png snippet using \code{latex} and \code{dvipng} from an stable
object; both functions are required to be installed for this to work.
The resultant \code{png} file is saved on disk and the relative path to the file
is returned. \code{st2png()} is an alias to \code{st_as_png()}.
}
\examples{
\dontrun{
Sys.which("latex")
Sys.which("dvipng")

tab <- stable(stdata())
st_aspng(tab)
}


}
\seealso{
\code{\link[=st_aspdf]{st_aspdf()}}, \code{\link[=st_as_image]{st_as_image()}}, \code{\link[=st_image_show]{st_image_show()}}
}
