% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_caption}
\alias{st_caption}
\title{Set table caption}
\usage{
st_caption(x, ..., short = NULL, short_repeat = TRUE, short_sep = NULL)
}
\arguments{
\item{x}{an stobject.}

\item{...}{character lines to form a table caption; see details.}

\item{short}{an abbreviated form of the caption to be used in a list of
tables.}

\item{short_repeat}{logical; if a short caption is provided, it will be
repeated at the start of the main caption.}

\item{short_sep}{a character sequence used to separate the short title
with the main caption when a short title is specified.}
}
\description{
Set table caption
}
\details{
A short title can also be specified by including it in brackets (\verb{[]}) as
the first text in \code{...}. See examples.
}
\examples{
tab <- st_new(stdata())
tab <- st_caption(tab, "[Full covariate model estimates]. Run number 101.")
tab$caption

}
