% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_caption}
\alias{st_caption}
\title{Set table caption}
\usage{
st_caption(x, text, ...)
}
\arguments{
\item{x}{an stobject.}

\item{text}{caption text to be parsed.}

\item{...}{additional arguments passed to \code{\link[=as.caption]{as.caption()}}.}
}
\description{
Set table caption
}
\details{
A short title can also be specified by including it in brackets (\verb{[]}) as
the first characters in \code{text}. See examples here as well as in
\code{\link[=as.caption]{as.caption()}}.
}
\examples{
tab <- st_new(stdata())
tab <- st_caption(tab, "[Full covariate model estimates]. Run number 101.")
tab$caption

text <- stable(tab)
text

}
