% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_clear_reps}
\alias{st_clear_reps}
\title{Add clear rep information to st object}
\usage{
st_clear_reps(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{quoted or unquoted column names passed to \code{\link[=stable]{stable()}} as
\code{clear_reps}}
}
\description{
See the \code{clear_reps} argument to \code{\link[=stable]{stable()}}.
}
