% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_clone}
\alias{st_clone}
\title{Clone an stobject}
\usage{
st_clone(x)
}
\arguments{
\item{x}{an stobject object}
}
\value{
A (deep) copy of \code{x}.
}
\description{
Clone an stobject
}
\examples{
x <- st_new(stdata)
y <- st_clone(x)

x
y

}
