% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_files}
\alias{st_files}
\title{Add file name information to st object}
\usage{
st_files(x, r = NULL, output = NULL)
}
\arguments{
\item{x}{an stobject}

\item{r}{set \code{r_file}, passed to \code{\link[=stable]{stable()}}}

\item{output}{set \code{output_file}, passed to \code{\link[=stable]{stable()}}}
}
\description{
Add file name information to st object
}
