% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_hline}
\alias{st_hline}
\title{Add hline information to st object}
\usage{
st_hline(
  x,
  pattern = NULL,
  cols = names(x$data),
  n = 1,
  at = NULL,
  from = NULL,
  nudge = 0
)
}
\arguments{
\item{x}{and stobject}

\item{pattern}{a regular expression to find rows where an \code{hline} will be
placed; passed to \code{\link[stringr:str_detect]{stringr::str_detect()}}}

\item{cols}{data columns to scan using \code{pattern}}

\item{n}{number of \code{hlines} to add when a hit is found}

\item{at}{logical or integer locations for hline passed to \code{\link[=stable]{stable()}} as
\code{hline_at}}

\item{from}{character column name used to divide a table; passed to
\code{\link[=stable]{stable()}} as \code{hline_from}}

\item{nudge}{push an hline down or up in the table; only applies to
indices found from the \code{at} or \code{pattern} arguments}
}
\description{
See the \code{hline_at} and \code{hline_from} arguments to \code{\link[=stable]{stable()}}.
}
