% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_new}
\alias{st_new}
\alias{st_new.data.frame}
\alias{st_new.pmtable}
\alias{st_data}
\title{Create an st object}
\usage{
st_new(x, ...)

\method{st_new}{data.frame}(x, ...)

\method{st_new}{pmtable}(x, ...)

st_data(data, ...)
}
\arguments{
\item{x}{either a data frame or an object of class \code{pmtable}; see details.}

\item{...}{additional arguments which will eventually get passed to the
table render function (e.g. \code{\link[=stable]{stable()}} or \code{\link[=stable_long]{stable_long()}}).}
}
\value{
And object with class \code{stobject} which can get piped to other functions. The
\code{pmtable} method returns an object that also has class \code{ptobject}.
}
\description{
The st object will collect various configuration settings and pass those
to \code{\link[=stable]{stable()}} when the object is passed to \code{\link[=st_make]{st_make()}}.
}
\details{
Methods are included for \code{data.frame} and \code{pmtable}, an object that comes
from one of the data summary functions (e.g. \code{\link[=pt_cont_wide]{pt_cont_wide()}}, or
\code{\link[=pt_cat_long]{pt_cat_long()}} or \code{\link[=pt_demographics]{pt_demographics()}}).

If using the data frame method, the user should filter or subset so that
the data (\code{x}) contains exactly the rows (and columns) to be processed;
pmtables will not add or remove rows prior to processing \code{x}.
}
\examples{
ob <- st_new(ptdata())
ob <- st_data(ptdata())

ob <- st_new(pt_data_inventory(pmt_obs))

}
