% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_new}
\alias{st_new}
\alias{st_data}
\title{Create an st object}
\usage{
st_new(data, ...)

st_data(data, ...)
}
\arguments{
\item{data}{the data frame to pass to \code{\link[=stable]{stable()}}; the user should filter
or subset so that \code{data} contains exactly the rows (and columns) to be
processed; pmtables will not add or remove rows prior to processing \code{data}}

\item{...}{additional arguments passed to \code{\link[=stable]{stable()}}}
}
\description{
The st object will collect various configuration settings and pass those
to \code{\link[=stable]{stable()}} when the object is passed to \code{\link[=st_make]{st_make()}}.
}
\examples{
ob <- st_new(ptdata())
ob <- st_data(ptdata())

}
