% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_noteconf}
\alias{st_noteconf}
\title{Add note config information to st object}
\usage{
st_noteconf(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{named arguments passed to \code{\link[=noteconf]{noteconf()}}}
}
\description{
See the \code{note_config} argument passed to \code{\link[=stable]{stable()}} and then to
\code{\link[=tab_notes]{tab_notes()}}.
}
\examples{
library(dplyr)

ob <- st_new(ptdata())

ob \%>\%
  st_notes("ALB: albumin (g/dL)") \%>\%
  st_noteconf(type = "minipage") \%>\%
  st_make()

}
