% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_notes}
\alias{st_notes}
\title{Add note information to st object}
\usage{
st_notes(x, ..., config = NULL)
}
\arguments{
\item{x}{an stobject}

\item{...}{table notes}

\item{config}{named list of arguments for \code{\link[=noteconf]{noteconf()}}}
}
\description{
Add note information to st object
}
