% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_panel}
\alias{st_panel}
\title{Add panel information to st object}
\usage{
st_panel(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{passed to \code{\link[=rowpanel]{rowpanel()}}}
}
\description{
See the \code{panel} argument to \code{\link[=stable]{stable()}}.
}
\examples{
library(dplyr)

ob <- st_new(ptdata())

ob \%>\% st_panel("STUDY") \%>\% st_make()

}
