% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_rename}
\alias{st_rename}
\title{Add column rename information to st object}
\usage{
st_rename(x, ..., .list = NULL)
}
\arguments{
\item{x}{an stobject}

\item{...}{column rename items in \code{new-name = old-name} format; passed
to \code{\link[=stable]{stable()}} as \code{cols_rename}}

\item{.list}{allows for the use of a list as an input alternative to \code{...}.}
}
\description{
See the \code{cols_rename} argument to \code{\link[=stable]{stable()}}. This function can be called
multiple times and will accumulate \code{cols_rename} data.
}
\examples{
library(dplyr)

st_new(ptdata()) \%>\% st_rename(Weight = WT) \%>\% stable()

}
