% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_select}
\alias{st_select}
\alias{st_mutate}
\title{Filter, select, or mutate data}
\usage{
st_select(x, ...)

st_mutate(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{passed to \code{\link[dplyr:select]{dplyr::select()}}, or \code{\link[dplyr:mutate]{dplyr::mutate()}}}
}
\description{
Filter, select, or mutate data
}
