% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_space}
\alias{st_space}
\title{Add row and column spacing information to st object}
\usage{
st_space(x, row = 1.5, col = 5)
}
\arguments{
\item{x}{an stobject}

\item{row}{set \code{row_space}, passed to \code{\link[=stable]{stable()}}}

\item{col}{set \code{col_space}, passed to \code{\link[=stable]{stable()}}}
}
\description{
See the \code{row_space} and \code{col_space} arguments to \code{\link[=stable]{stable()}}.
}
\examples{
library(dplyr)

ob <- st_new(ptdata())

ob \%>\% st_space(row = 1) \%>\% st_make()

}
