% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_span}
\alias{st_span}
\title{Add column spanner to st object}
\usage{
st_span(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{passed to \code{\link[=colgroup]{colgroup()}}}
}
\description{
Add column spanner to st object
}
