% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_span}
\alias{st_span}
\title{Add column spanner to st object}
\usage{
st_span(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{passed to \code{\link[=colgroup]{colgroup()}}}
}
\description{
See the \code{span} argument to \code{\link[=stable]{stable()}}. This function can be called multiple
times and will accumulate \code{span} data.
}
\examples{
library(dplyr)

ob <- st_new(ptdata())

ob \%>\% st_span("Covariates", WT:ALB) \%>\% st_make()

}
