% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_span_split}
\alias{st_span_split}
\title{Add column split spanner information to st object}
\usage{
st_span_split(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{passed to \code{\link[=colsplit]{colsplit()}}}
}
\description{
See the \code{span_split} argument to \code{\link[=stable]{stable()}}.
}
\examples{
library(dplyr)

file <- system.file("datasets", "with-dots.RDS", package = "pmtables")

data <- readRDS(file)

data \%>\% st_span_split('.') \%>\% st_make()

}
