% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-object.R
\name{st_sumrow}
\alias{st_sumrow}
\title{Add summary row information to st object}
\usage{
st_sumrow(x, ...)
}
\arguments{
\item{x}{an stobject}

\item{...}{passed to \code{\link[=sumrow]{sumrow()}}}
}
\description{
Add summary row information to st object
}
