% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_to_standalone}
\alias{st_to_standalone}
\title{Convert stable text to a standalone snippet}
\usage{
st_to_standalone(
  text,
  stem,
  dir,
  font = c("helvetica", "roboto", "utopia"),
  command = "latex",
  textwidth = 6.5,
  border = "0.2cm 1cm"
)
}
\arguments{
\item{text}{character vector of table text.}

\item{stem}{used to build intermediate and output file names.}

\item{dir}{directory for building the pdf file.}

\item{font}{the font to use; alternative values include \code{roboto} and
\code{utopia}; passed to \code{\link[=st_to_standalone]{st_to_standalone()}}.}

\item{command}{pass \code{pdflatex} when building a \code{pdf} file or \code{latex} when
building \code{png}.}

\item{textwidth}{used to set the width of the page.}

\item{border}{passed as an option to \code{standalone} latex output type; this
can be one, two or four space-separated elements, each formatted as
\code{"<number><unit>"} (e.g. "0.2cm"); pass one element to set the same border on
all sides; two elements to set the border on left/right (first) and
top/bottom (second); pass four elements to have separate borders for
the left, bottom, right and top (see the documentation for the \code{standalone}
latex package).}
}
\description{
Convert stable text to a standalone snippet
}
\keyword{internal}
