% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-standalone.R
\name{st_to_standalone}
\alias{st_to_standalone}
\title{Convert stable text to a standalone snippet}
\usage{
st_to_standalone(
  text,
  stem,
  dir,
  font = c("helvetica", "roboto", "utopia"),
  command = "latex",
  textwidth = 6.5,
  border = "0.2cm 1cm"
)
}
\arguments{
\item{text}{character vector of table text.}

\item{stem}{used to build intermediate and output file names.}

\item{dir}{directory for building the pdf file.}

\item{font}{the font to use; alternative values include \code{roboto} and
\code{utopia}; passed to \code{\link[=st_to_standalone]{st_to_standalone()}}.}

\item{command}{pass \code{pdflatex} when building a \code{pdf} file or \code{latex} when
building \code{png}.}

\item{textwidth}{the page width (in inches) when building with \code{pdflatex};
passed to \code{\link[=st_to_standalone]{st_to_standalone()}}; see details.}

\item{border}{passed as an option to \code{standalone} latex output type; see
details.}
}
\description{
Convert stable text to a standalone snippet
}
\keyword{internal}
