% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-functions.R
\name{st_wrap}
\alias{st_wrap}
\alias{pt_wrap}
\title{Wrap stable output in table environment}
\usage{
st_wrap(x, con = NULL, center = TRUE, caption = NULL)

pt_wrap(x, con = NULL, center = TRUE, caption = NULL)
}
\arguments{
\item{x}{tabular text}

\item{con}{where to write the output}

\item{center}{if \code{TRUE}, center the table}

\item{caption}{the table caption}
}
\description{
Wrap stable output in table environment
}
