% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{st_wrap}
\alias{st_wrap}
\alias{st_wrap.default}
\alias{st_wrap.stable_long}
\alias{pt_wrap}
\alias{st_asis}
\alias{st_latex}
\title{Wrap stable output in table environment}
\usage{
st_wrap(x, ...)

\method{st_wrap}{default}(
  x,
  con = stdout(),
  table = TRUE,
  center = TRUE,
  landscape = is_lscape(x),
  caption = NULL,
  short = NULL,
  float = c("H", "!ht"),
  context = c("rmd", "tex"),
  asis = FALSE,
  ...
)

\method{st_wrap}{stable_long}(x, table = FALSE, ...)

pt_wrap(x, ...)

st_asis(x, ..., asis = NULL, con = NULL)

st_latex(x, con = stdout(), center = TRUE, context = c("rmd", "tex"))
}
\arguments{
\item{x}{tabular text}

\item{...}{not used}

\item{con}{where to write the output}

\item{table}{if \code{TRUE}, the code is wrapped in latex table environment}

\item{center}{if \code{TRUE}, center the table}

\item{landscape}{if \code{TRUE} render the table in landscape mode}

\item{caption}{the long table description}

\item{short}{the short table description}

\item{float}{the float specifier to if a \code{table} environment is used; change
this to \code{!ht} if the float package cannot be loaded for some reason}

\item{context}{if \code{rmd}, then the code is enclosed in a pandoc \code{latex} fenced
code block; if \code{tex}, then the fencing is omitted}

\item{asis}{if \code{TRUE}, the wrapped table is processed with
\code{\link[knitr:asis_output]{knitr::asis_output()}} and returned}
}
\description{
Wrap stable output in table environment
}
\seealso{
\code{\link[=st_use_knit_deps]{st_use_knit_deps()}}
}
