% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-stable.R
\name{stable}
\alias{stable}
\title{Create tabular output from an R data frame}
\usage{
stable(
  data,
  align = cols_left(),
  panel = rowpanel(col = NULL),
  units = NULL,
  clear_reps = NULL,
  rm_dups = NULL,
  span = NULL,
  span_split = NULL,
  notes = NULL,
  hline_at = NULL,
  hline_from = NULL,
  sumrows = NULL,
  bold_cols = missing(panel),
  col_rename = NULL,
  col_replace = NULL,
  row_space = 1.4,
  col_space = 5,
  fontsize = NULL,
  prime_fun = tab_prime,
  escape_fun = tab_escape,
  note_config = noteconf(type = "tpt"),
  r_file = NULL,
  r_file_label = getOption("r.file.label", "Source code:"),
  output_file = NULL,
  output_file_label = getOption("out.file.label", "Source file: ")
)
}
\arguments{
\item{data}{a data.frame to convert to tabular table}

\item{align}{an alignment object created by \link{cols_align}, \link{cols_left},
\link{cols_center}, or \link{cols_right}}

\item{panel}{character column name to use to section the table; sections will
be created from unique values of \code{data[[panel]]}}

\item{units}{a named list with unit information; names should correspond to
columns in the data frame}

\item{clear_reps}{character vector of column names where duplicate values will
be made blank (overwritten with \code{""})}

\item{rm_dups}{use \code{clear_reps}}

\item{span}{a list of objects created with \link{colgroup}}

\item{span_split}{not implemented at this time}

\item{notes}{a character vector of notes to include at the foot of the table;
use \code{r_file} and \code{output_file} for source code and output file annotations}

\item{hline_at}{logical or integer vector specifying rows above which an
\verb{\\hline} will be placed}

\item{hline_from}{a character column name from which to separate the table
with \verb{\\hline}; non-duplicated values of \code{hline_from} will be used to create
the split}

\item{sumrows}{an object created with \link{sumrow}; identifies summary rows
and adds styling}

\item{bold_cols}{if \code{TRUE}, table column names are rendered with bold font}

\item{col_rename}{a \code{name = value} character vector to translate column names
to table names}

\item{col_replace}{a character vector with the same length as the number of
output table columns; use this to completely replace the names (as opposed
to one by on editing with \code{col_rename})}

\item{row_space}{relative increase or decrease spacing between rows; use
\verb{row_space > <default>} to increase}

\item{col_space}{absolute column spacing amount (\code{pt})}

\item{fontsize}{for the table (e.g. \code{normalsize}, \code{small}, \code{scriptsize}, etc)}

\item{prime_fun}{function to prime the data frame to be converted to tabular}

\item{escape_fun}{a function passed to \code{prime_fun} that will sanitize column
data}

\item{note_config}{a \code{\link[=noteconf]{noteconf()}} object used to configure how table notes
are displayed}

\item{r_file}{the name of the R file containg code to generate the table; the
file name will be included in the notes in the table footer}

\item{r_file_label}{prefix text for \code{r_file} note}

\item{output_file}{the name of the output file where the table text will be
saved; the file name will be included in the notes in the table footer}

\item{output_file_label}{prefix text for \code{output_file} note}
}
\description{
Create tabular output from an R data frame
}
