% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-stable.R
\name{stable}
\alias{stable}
\alias{stable.data.frame}
\alias{stable.pmtable}
\title{Create tabular output from an R data frame}
\usage{
stable(data, ...)

\method{stable}{data.frame}(
  data,
  align = cols_left(),
  panel = rowpanel(col = NULL),
  span = NULL,
  notes = NULL,
  sumrows = NULL,
  units = NULL,
  sizes = tab_size(),
  escape_fun = tab_escape,
  inspect = FALSE,
  ...
)

\method{stable}{pmtable}(data, ...)
}
\arguments{
\item{data}{a data.frame to convert to tabular table; see also \code{\link[=st_new]{st_new()}}}

\item{...}{passed to other functions: \code{\link[=tab_hlines]{tab_hlines()}}, \code{\link[=tab_spanners]{tab_spanners()}},
\code{\link[=tab_notes]{tab_notes()}}, \code{\link[=tab_cols]{tab_cols()}}, \code{\link[=tab_clear_reps]{tab_clear_reps()}} and \code{\link[=make_tabular]{make_tabular()}}}

\item{align}{an alignment object created by \code{\link[=cols_align]{cols_align()}}, \code{\link[=cols_left]{cols_left()}},
\code{\link[=cols_center]{cols_center()}}, or \code{\link[=cols_right]{cols_right()}}; see also \code{\link[=st_align]{st_align()}}}

\item{panel}{character column name to use to section the table; sections will
be created from unique values of \code{data[[panel]]}; see also \code{\link[=st_panel]{st_panel()}}}

\item{span}{a list of objects created with \code{\link[=colgroup]{colgroup()}}; ; see also \code{\link[=st_span]{st_span()}}}

\item{notes}{a character vector of notes to include at the foot of the table;
use \code{r_file} and \code{output_file} for source code and output file annotations;
see \code{\link[=tab_notes]{tab_notes()}} for arguments to pass in order to configure the way notes
appear in the output; see also \code{\link[=st_notes]{st_notes()}}}

\item{sumrows}{an object created with \code{\link[=sumrow]{sumrow()}}; identifies summary rows
and adds styling; see also \code{\link[=st_sumrow]{st_sumrow()}}}

\item{units}{a named list with unit information; names should correspond to
columns in the data frame}

\item{sizes}{an object returned from \code{\link[=tab_size]{tab_size()}}}

\item{escape_fun}{a function passed to \code{prime_fun} that will sanitize column
data}

\item{inspect}{if \code{TRUE}, extra information is attached to the output
as an attribute called \code{stable_data}; see \code{\link[=get_stable_data]{get_stable_data()}}}
}
\description{
Create tabular output from an R data frame
}
\examples{
data <- ptdata()

a <- stable(data, r_file = "example.R", output_file = "output.tex")

b <- stable(data, panel = "STUDYf")

c <- stable(data, span = colgroup("Covariates", STUDYf:ALB))

}
