% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-long.R
\name{stable_long}
\alias{stable_long}
\title{Create longtable output from an R data frame}
\usage{
stable_long(..., debug_data = TRUE)
}
\arguments{
\item{...}{passed to \code{\link[=stable]{stable()}}}

\item{debug_data}{fixed to \code{TRUE} and passed to \code{\link[=stable]{stable()}}}
}
\description{
Create longtable output from an R data frame
}
