% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-long.R
\name{stable_long}
\alias{stable_long}
\title{Create longtable output from an R data frame}
\usage{
stable_long(
  data,
  note_config = noteconf(type = "minipage"),
  inspect = FALSE,
  lt_cap_macro = "",
  lt_cap_text = "",
  lt_cap_short = "",
  lt_cap_label = "",
  lt_continue = "\\\\footnotesize{continued on next page}",
  ...
)
}
\arguments{
\item{data}{data frame passed to \code{\link[=stable]{stable()}}; the user should filter
or subset so that \code{data} contains exactly the rows (and columns) to be
processed; pmtables will not add or remove rows prior to processing \code{data}}

\item{note_config}{a \code{\link[=noteconf]{noteconf()}} object used to configure how table notes
are displayed; ; see also \code{\link[=st_noteconf]{st_noteconf()}}}

\item{inspect}{fixed to \code{TRUE} and passed to \code{\link[=stable]{stable()}}}

\item{lt_cap_macro}{the name of a macro that will hold caption text; to not lead with
\verb{\\\\} - this will be added for you}

\item{lt_cap_text}{full caption text, appearing where the table is rendered}

\item{lt_cap_short}{short caption text, appearing in the list of tables}

\item{lt_cap_label}{table label for use in latex document}

\item{lt_continue}{longtable continuation message}

\item{...}{passed to \code{\link[=stable]{stable()}}}
}
\description{
Create longtable output from an R data frame
}
