% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-utils.R
\name{stable_save}
\alias{stable_save}
\title{Save output from stable}
\usage{
stable_save(x, file = attr(x, "stable_file"), dir = NULL)
}
\arguments{
\item{x}{a table formatted with \code{\link[=stable]{stable()}}}

\item{file}{the file}

\item{dir}{the directory where the file is to be saved}
}
\description{
Save output from stable
}
