% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-cols.R
\name{tab_cols}
\alias{tab_cols}
\title{Form table column names}
\usage{
tab_cols(
  cols,
  col_replace = NULL,
  col_rename = NULL,
  col_blank = NULL,
  col_split = NULL,
  col_bold = FALSE,
  pull_back = NULL,
  ...
)
}
\arguments{
\item{cols}{the starting set of table column names}

\item{col_replace}{a character vector with the same length as the number of
output table columns; use this to completely replace the names (as opposed
to one by on editing with \code{col_rename})}

\item{col_rename}{a \code{name = value} character vector to translate column names
to table names; ; see also \code{\link[=st_rename]{st_rename()}}}

\item{col_blank}{a character vector of column names that will not be printed
in the table header; see also \code{\link[=st_blank]{st_blank()}}}

\item{col_split}{a string that is used to split column labels into tag
(on the left) or name (on the right); if supplied, then \code{col_split} will be
used to remove the tag; for example, a column named \code{x.WT} would be renamed
\code{WT} if \code{col_split} was set to \code{.}}

\item{col_bold}{if \code{TRUE}, table column names are rendered with bold font}

\item{pull_back}{pull unit row back, slightly closer to table columns labels}

\item{...}{not used}
}
\description{
Form table column names
}
