% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-notes.R
\name{tab_notes}
\alias{tab_notes}
\title{Form table notes}
\usage{
tab_notes(
  notes = character(0),
  escape_fun = tab_escape,
  mask_bracket = c("both", "left", "right", "none"),
  note_config = noteconf(type = "tpt"),
  r_file = getOption("mrg.script", NULL),
  r_file_label = "Source code: ",
  output_file = NULL,
  output_file_label = "Source file: ",
  output_dir = getOption("pmtables.dir"),
  path.type = getOption("pmtables.path.type", "none"),
  ...
)
}
\arguments{
\item{notes}{a character vector of notes to include at the foot of the table;
use \code{r_file} and \code{output_file} for source code and output file annotations;
see \code{\link[=tab_notes]{tab_notes()}} for arguments to pass in order to configure the way notes
appear in the output; see also \code{\link[=st_notes]{st_notes()}}}

\item{escape_fun}{a function passed to \code{prime_fun} that will sanitize column
data}

\item{mask_bracket}{mask (or substitute) left and/or right brackets with \verb{\\lbrack}
and/or \verb{\\rbrack}, respectively.}

\item{note_config}{a \code{\link[=noteconf]{noteconf()}} object used to configure how table notes
are displayed; see also \code{\link[=st_noteconf]{st_noteconf()}}.}

\item{r_file}{the name of the R file containing code to generate the table;
the file name will be included in the notes in the table footer; see also
\code{\link[=st_files]{st_files()}}.}

\item{r_file_label}{a prefix for \code{r_file}.}

\item{output_file}{the name of the output file where the table text will be
saved; the file name will be included in the notes in the table footer; see
also \code{\link[=st_files]{st_files()}}.}

\item{output_file_label}{a prefix for \code{output_file}.}

\item{output_dir}{directory location where output \code{.tex} file is saved;
defaults to working directory.}

\item{path.type}{whether to include the path to the output file in the
table notes and how to format it; options include "none", "proj", and
"raw"; see \code{\link[=format_table_path]{format_table_path()}}.}

\item{...}{not used.}
}
\description{
Form table notes
}
\examples{
tab <- stable(
  stdata(),
  note_config = noteconf(type = "mini")
)

}
