% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sizes.R
\name{tab_size}
\alias{tab_size}
\title{Adjustments for row padding, column padding, and font size}
\usage{
tab_size(row = 1.3, col = 5, font = NULL, header_row = -0.4 * row, lt_row = 0)
}
\arguments{
\item{row}{relative increase or decrease spacing between rows; use
\verb{row_space > <default>} to increase; this parameter applies to both long tables
ans tabular tables; see also \code{\link[=st_space]{st_space()}}}

\item{col}{absolute column spacing amount (\code{pt});  this parameter applies to both
long tables and tabular tables; see also \code{\link[=st_space]{st_space()}}}

\item{font}{for the table (e.g. \code{normalsize}, \code{small}, \code{scriptsize}, etc)}

\item{header_row}{extra space between rows in the header
section of the table (column names and units); this should be a negative
number if you wish to compress header rows; the default is \code{-0.4*row};
recommend fixing to \code{-0.2} when \code{row} is 1 and smaller if \code{row} is less
than 1}

\item{lt_row}{extra row space for longtables; this is only needed for
extra control over row height; this can be a positive or negative
number and a value of 0 indicates to neither add or subtract row space}
}
\description{
Adjustments for row padding, column padding, and font size
}
